Candidates' Color Priority

Classification of candidates according to waiting list time on dialysis' quartiles and two cPRA cutoff values.

Usage

cp(data = candidates, q2 = 60, q3 = 100, cPRA1 = 50, cPRA2 = 85)

Arguments

data

A data frame with information for candidates' waiting list.

q2

A numerical value for the median of candidates' waiting list (0 - 120).

q3

A numerical value for the third quartile of candidates' waiting list (0 - 120).

cPRA1

A numerical value (0 - 100) for the lower cPRA cutoff.

cPRA2

A numerical value (0 - 100) for the higher cPRA cutoff. cPRA2 must be greater than cPRA1.

Value

A data frame with a new column 'cp' (color priority)

Examples

cp(data = candidates, q2 = 60, q3 = 100, cPRA1 = 50, cPRA2 = 85)[1:5,]
#>   ID bg A1 A2 B1 B2 DR1 DR2 age dialysis cPRA urgent    cp
#> 1  1  A  2 24 15 44   1   4  53       59    0      0 Green
#> 2  2  O  1  2 57 58   4   7  41       35   16      0 Green
#> 3  3  O 25 29 18 44   4   7  72       53    0      0 Green
#> 4  4  O  2 31 14 44   1  12  48       50    0      0 Green
#> 5  5  O  3 24  7 35   1  15  62       39    0      0 Green