Candidates' selection according to any algorithm for multiple donors

Ordering of waitlisted candidates for a given donor and according to any algorithm.

Usage

donor_recipient_pairs(
  df.donors = donors,
  df.candidates = candidates,
  df.abs = cabs,
  algorithm = lima,
  n = 2,
  check.validity = TRUE,
  ...
)

Arguments

df.donors

A data frame containing demographics and medical information for a poll of donors. For uk algorithm must have respective columns.

df.candidates

A data frame containing demographics and medical information for a group of waitlisted transplant candidates. For uk algorithm must have respective columns.

df.abs

A data frame with candidates' antibodies.

algorithm

The name of the function to use. Valid options are: lima, et, pt, uk (without quotation)

n

A positive integer to slice the first candidates.

check.validity

Logical to decide whether to validate input.

...

all the parameters used on the function algorithm

Value

A list with the number of elements equal to the number of rows on donors data frame. Each element have a data frame with select candidates by donor.

Examples

donor_recipient_pairs(df.donors = donors,
df.candidates = candidates,
df.abs = cabs,
algorithm = lima,
n = 2,
check.validity = TRUE)
#> [[1]]
#>     ID bg A1 A2 B1 B2 DR1 DR2 mmA mmB mmDR mmHLA age donor_age dialysis cPRA
#> 1:  81  A  2 23 15 49  15   4   1   1    1     3  44        46       91    0
#> 2: 243  A  2 24  7 49   4   4   1   1    1     3  22        46       73    0
#>       HI     cp SP
#> 1: FALSE Yellow  2
#> 2: FALSE Yellow  2
#> 
#> [[2]]
#>     ID bg A1 A2 B1 B2 DR1 DR2 mmA mmB mmDR mmHLA age donor_age dialysis cPRA
#> 1: 317  O  1 68 51 58  13   9   1   2    2     5  31        55      113    0
#> 2: 112  O  2 29 44 44   4  12   2   2    2     6  58        55      104    0
#>       HI     cp SP
#> 1: FALSE Orange  2
#> 2: FALSE Orange  2
#> 
#> [[3]]
#>     ID bg A1 A2 B1 B2 DR1 DR2 mmA mmB mmDR mmHLA age donor_age dialysis cPRA
#> 1: 317  O  1 68 51 58  13   9   2   2    1     5  31        32      113    0
#> 2: 487  O  2 33 15 27  11   7   2   1    2     5  55        32      103    0
#>       HI     cp SP
#> 1: FALSE Orange  2
#> 2: FALSE Orange  2
#> 
#> [[4]]
#>     ID bg A1 A2 B1 B2 DR1 DR2 mmA mmB mmDR mmHLA age donor_age dialysis cPRA
#> 1: 464  A  1  3 35 57  13   7   1   2    1     4  52        35       48   88
#> 2: 129  A  2 24 40 45  14  10   2   2    2     6  59        35       75   90
#>      HI     cp SP
#> 1: TRUE Orange  2
#> 2: TRUE Orange  2
#> 
#> [[5]]
#>     ID bg A1 A2 B1 B2 DR1 DR2 mmA mmB mmDR mmHLA age donor_age dialysis cPRA
#> 1: 215  A  2 25 50 51  11   7   1   2    1     4  68        73       66    0
#> 2: 239  A 11 32 35 44   4   7   1   2    1     4  66        73       66    0
#>       HI     cp SP
#> 1: FALSE Yellow  1
#> 2: FALSE Yellow  1
#> 
#> [[6]]
#>     ID bg A1 A2 B1 B2 DR1 DR2 mmA mmB mmDR mmHLA age donor_age dialysis cPRA
#> 1: 129  A  2 24 40 45  14  10   1   2    2     5  59        41       75   90
#> 2:  81  A  2 23 15 49  15   4   1   1    2     4  44        41       91    0
#>       HI     cp SP
#> 1:  TRUE Orange  2
#> 2: FALSE Yellow  2
#> 
#> [[7]]
#>     ID bg A1 A2 B1 B2 DR1 DR2 mmA mmB mmDR mmHLA age donor_age dialysis cPRA
#> 1:  16  O  3 24  8 18   3  13   1   1    1     3  66        75       60    0
#> 2: 324  O  1 11  8 35   3  13   1   2    1     4  69        75       81    0
#>       HI     cp SP
#> 1: FALSE Yellow  1
#> 2: FALSE Yellow  1
#> 
#> [[8]]
#>    ID bg A1 A2 B1 B2 DR1 DR2 mmA mmB mmDR mmHLA age donor_age dialysis cPRA
#> 1: 65 AB  2  2 35 51  11   4   1   2    2     5  62        64       63    0
#> 2: 74 AB  2 11 14 57  11   7   1   2    1     4  39        64       52    0
#>       HI     cp SP
#> 1: FALSE Yellow  2
#> 2: FALSE  Green  2
#> 
#> [[9]]
#>     ID bg A1 A2 B1 B2 DR1 DR2 mmA mmB mmDR mmHLA age donor_age dialysis cPRA
#> 1: 445  A  2 66 49 51   4  11   0   1    1     2  40        54       81   98
#> 2: 143  A  2 24  7 44  15   4   0   1    1     2  20        54       73   86
#>      HI     cp SP
#> 1: TRUE Orange  2
#> 2: TRUE Orange  2
#> 
#> [[10]]
#>     ID bg A1 A2 B1 B2 DR1 DR2 mmA mmB mmDR mmHLA age donor_age dialysis cPRA
#> 1: 143  A  2 24  7 44  15   4   1   1    1     3  20        40       73   86
#> 2: 386  A  1  3 44 51   4   7   1   0    0     1  47        40       71    0
#>       HI     cp SP
#> 1:  TRUE Orange  2
#> 2: FALSE Yellow  2
#> 
#> [[11]]
#>     ID bg A1 A2 B1 B2 DR1 DR2 mmA mmB mmDR mmHLA age donor_age dialysis cPRA
#> 1: 212  A 23 32 14 44   7   7   1   1    1     3  53        43       84    0
#> 2: 286  A  3 29 35 44   1   4   1   0    2     3  25        43       72    0
#>       HI     cp SP
#> 1: FALSE Yellow  2
#> 2: FALSE Yellow  2
#> 
#> [[12]]
#>     ID bg A1 A2 B1 B2 DR1 DR2 mmA mmB mmDR mmHLA age donor_age dialysis cPRA
#> 1: 242  O  2 11 18 35   1  11   1   0    1     2  54        50       81   94
#> 2: 487  O  2 33 15 27  11   7   1   1    1     3  55        50      103    0
#>       HI     cp SP
#> 1:  TRUE Orange  2
#> 2: FALSE Orange  2
#> 
#> [[13]]
#>     ID bg A1 A2 B1 B2 DR1 DR2 mmA mmB mmDR mmHLA age donor_age dialysis cPRA
#> 1: 124  A  2 24 41 52  15   3   0   2    1     3  30        32       27   88
#> 2: 174  A  1 24 35 39   1  13   1   1    1     3  48        32       75    0
#>       HI     cp SP
#> 1:  TRUE Orange  2
#> 2: FALSE Yellow  2
#> 
#> [[14]]
#>     ID bg A1 A2 B1 B2 DR1 DR2 mmA mmB mmDR mmHLA age donor_age dialysis cPRA
#> 1: 317  O  1 68 51 58  13   9   2   1    1     4  31        51      113    0
#> 2: 112  O  2 29 44 44   4  12   1   2    2     5  58        51      104    0
#>       HI     cp SP
#> 1: FALSE Orange  2
#> 2: FALSE Orange  2
#> 
#> [[15]]
#>     ID bg A1 A2 B1 B2 DR1 DR2 mmA mmB mmDR mmHLA age donor_age dialysis cPRA
#> 1: 205  A  2  3 35 50   3  11   1   2    1     4  43        38       62   96
#> 2: 298  A  3 33 51 51  15  16   1   1    1     3  31        38       73    0
#>       HI     cp SP
#> 1:  TRUE Orange  2
#> 2: FALSE Yellow  2
#> 
#> [[16]]
#>     ID bg A1 A2 B1 B2 DR1 DR2 mmA mmB mmDR mmHLA age donor_age dialysis cPRA
#> 1: 205  A  2  3 35 50   3  11   0   1    2     3  43        55       62   96
#> 2: 353  A  2 29  7 51   1  13   1   1    0     2  43        55       61    0
#>       HI     cp SP
#> 1:  TRUE Orange  2
#> 2: FALSE Yellow  2
#> 
#> [[17]]
#>     ID bg A1 A2 B1 B2 DR1 DR2 mmA mmB mmDR mmHLA age donor_age dialysis cPRA
#> 1: 317  O  1 68 51 58  13   9   1   1    1     3  31        52      113    0
#> 2: 487  O  2 33 15 27  11   7   2   2    1     5  55        52      103    0
#>       HI     cp SP
#> 1: FALSE Orange  2
#> 2: FALSE Orange  2
#> 
#> [[18]]
#>     ID bg A1 A2 B1 B2 DR1 DR2 mmA mmB mmDR mmHLA age donor_age dialysis cPRA
#> 1: 344  A  1  3 35 49   1   4   2   2    0     4  70        68       69    0
#> 2: 215  A  2 25 50 51  11   7   1   1    2     4  68        68       66    0
#>       HI     cp SP
#> 1: FALSE Yellow  1
#> 2: FALSE Yellow  1
#> 
#> [[19]]
#>     ID bg A1 A2 B1 B2 DR1 DR2 mmA mmB mmDR mmHLA age donor_age dialysis cPRA
#> 1:   6  B 23 33 14 44   7   7   0   2    1     3  51        46       76    0
#> 2: 328  B  3 33 38 51  11  12   1   2    2     5  27        46       70    0
#>       HI     cp SP
#> 1: FALSE Yellow  2
#> 2: FALSE Yellow  2
#> 
#> [[20]]
#>     ID bg A1 A2 B1 B2 DR1 DR2 mmA mmB mmDR mmHLA age donor_age dialysis cPRA
#> 1: 205  A  2  3 35 50   3  11   0   1    2     3  43        45       62   96
#> 2: 286  A  3 29 35 44   1   4   1   0    2     3  25        45       72    0
#>       HI     cp SP
#> 1:  TRUE Orange  2
#> 2: FALSE Yellow  2
#> 
#> [[21]]
#>     ID bg A1 A2 B1 B2 DR1 DR2 mmA mmB mmDR mmHLA age donor_age dialysis cPRA
#> 1: 212  A 23 32 14 44   7   7   1   2    1     4  53        45       84    0
#> 2:  22  A  1 11 51 57   4   7   2   1    1     4  50        45       71    0
#>       HI     cp SP
#> 1: FALSE Yellow  2
#> 2: FALSE Yellow  2
#> 
#> [[22]]
#>    ID bg A1 A2 B1 B2 DR1 DR2 mmA mmB mmDR mmHLA age donor_age dialysis cPRA
#> 1: 65 AB  2  2 35 51  11   4   1   2    1     4  62        55       63    0
#> 2: 74 AB  2 11 14 57  11   7   0   2    1     3  39        55       52    0
#>       HI     cp SP
#> 1: FALSE Yellow  2
#> 2: FALSE  Green  2
#> 
#> [[23]]
#>     ID bg A1 A2 B1 B2 DR1 DR2 mmA mmB mmDR mmHLA age donor_age dialysis cPRA
#> 1: 320  O  3 11  7 35  15  14   0   1    1     2  52        48      103   80
#> 2: 192  O  3 33 14 35   1  16   0   1    1     2  64        48       61   94
#>       HI     cp SP
#> 1: FALSE Orange  2
#> 2:  TRUE Orange  2
#> 
#> [[24]]
#>     ID bg A1 A2 B1 B2 DR1 DR2 mmA mmB mmDR mmHLA age donor_age dialysis cPRA
#> 1: 249  B  1  2  8 50   3   7   0   0    1     1  40        60       48   82
#> 2:  28  B  2 68  8 39   3  11   1   1    1     3  34        60       58   62
#>       HI     cp SP
#> 1: FALSE Yellow  2
#> 2: FALSE Yellow  2
#> 
#> [[25]]
#>     ID bg A1 A2 B1 B2 DR1 DR2 mmA mmB mmDR mmHLA age donor_age dialysis cPRA
#> 1: 112  O  2 29 44 44   4  12   1   2    1     4  58        56      104    0
#> 2: 487  O  2 33 15 27  11   7   1   2    1     4  55        56      103    0
#>       HI     cp SP
#> 1: FALSE Orange  2
#> 2: FALSE Orange  2
#> 
#> [[26]]
#>    ID bg A1 A2 B1 B2 DR1 DR2 mmA mmB mmDR mmHLA age donor_age dialysis cPRA
#> 1: 17  A  1 31  8 18   3  14   1   1    1     3  47        55       88    0
#> 2: 81  A  2 23 15 49  15   4   1   2    1     4  44        55       91    0
#>       HI     cp SP
#> 1: FALSE Yellow  2
#> 2: FALSE Yellow  2
#> 
#> [[27]]
#>    ID bg A1 A2 B1 B2 DR1 DR2 mmA mmB mmDR mmHLA age donor_age dialysis cPRA
#> 1: 22  A  1 11 51 57   4   7   1   1    1     3  50        52       71    0
#> 2: 17  A  1 31  8 18   3  14   2   1    1     4  47        52       88    0
#>       HI     cp SP
#> 1: FALSE Yellow  2
#> 2: FALSE Yellow  2
#> 
#> [[28]]
#>     ID bg A1 A2 B1 B2 DR1 DR2 mmA mmB mmDR mmHLA age donor_age dialysis cPRA
#> 1: 385  A  2 24 15 35   3  11   1   2    1     4  49        44      103   90
#> 2: 205  A  2  3 35 50   3  11   1   2    1     4  43        44       62   96
#>      HI     cp SP
#> 1: TRUE Orange  2
#> 2: TRUE Orange  2
#> 
#> [[29]]
#>     ID bg A1 A2 B1 B2 DR1 DR2 mmA mmB mmDR mmHLA age donor_age dialysis cPRA
#> 1: 487  O  2 33 15 27  11   7   2   2    0     4  55        52      103    0
#> 2: 112  O  2 29 44 44   4  12   2   1    2     5  58        52      104    0
#>       HI     cp SP
#> 1: FALSE Orange  2
#> 2: FALSE Orange  2
#> 
#> [[30]]
#>     ID bg A1 A2 B1 B2 DR1 DR2 mmA mmB mmDR mmHLA age donor_age dialysis cPRA
#> 1: 129  A  2 24 40 45  14  10   2   1    2     5  59        40       75   90
#> 2: 246  A  1 24 15 27  11  13   1   2    2     5  34        40       45   86
#>      HI     cp SP
#> 1: TRUE Orange  2
#> 2: TRUE Orange  2
#> 
#> [[31]]
#>     ID bg A1 A2 B1 B2 DR1 DR2 mmA mmB mmDR mmHLA age donor_age dialysis cPRA
#> 1: 340  A  1 11 18 35   1  13   1   0    1     2  56        38       63    0
#> 2:  27  A 11 29 35 44   1  11   1   1    0     2  29        38        2   52
#>       HI     cp SP
#> 1: FALSE Yellow  2
#> 2: FALSE Yellow  2
#> 
#> [[32]]
#>     ID bg A1 A2 B1 B2 DR1 DR2 mmA mmB mmDR mmHLA age donor_age dialysis cPRA
#> 1: 112  O  2 29 44 44   4  12   1   1    1     3  58        41      104    0
#> 2: 317  O  1 68 51 58  13   9   1   1    2     4  31        41      113    0
#>       HI     cp SP
#> 1: FALSE Orange  2
#> 2: FALSE Orange  2
#> 
#> [[33]]
#>     ID bg A1 A2 B1 B2 DR1 DR2 mmA mmB mmDR mmHLA age donor_age dialysis cPRA
#> 1: 397  O  1  2 39 57   7   8   0   2    1     3  68        65       83    0
#> 2: 470  O  2 29 44 44   4   7   0   2    1     3  65        65       77    0
#>       HI     cp SP
#> 1: FALSE Yellow  1
#> 2: FALSE Yellow  1
#> 
#> [[34]]
#>     ID bg A1 A2 B1 B2 DR1 DR2 mmA mmB mmDR mmHLA age donor_age dialysis cPRA
#> 1: 385  A  2 24 15 35   3  11   0   1    1     2  49        31      103   90
#> 2: 205  A  2  3 35 50   3  11   0   1    1     2  43        31       62   96
#>      HI     cp SP
#> 1: TRUE Orange  2
#> 2: TRUE Orange  2
#> 
#> [[35]]
#>     ID bg A1 A2 B1 B2 DR1 DR2 mmA mmB mmDR mmHLA age donor_age dialysis cPRA
#> 1: 112  O  2 29 44 44   4  12   1   1    1     3  58        42      104    0
#> 2: 317  O  1 68 51 58  13   9   2   2    2     6  31        42      113    0
#>       HI     cp SP
#> 1: FALSE Orange  2
#> 2: FALSE Orange  2
#> 
#> [[36]]
#>     ID bg A1 A2 B1 B2 DR1 DR2 mmA mmB mmDR mmHLA age donor_age dialysis cPRA
#> 1: 194  B  2 32 14 58   1  13   1   1    1     3  24        62       46   90
#> 2: 136  B 24 24 15 39   1   8   2   1    2     5  50        62       37   88
#>      HI     cp SP
#> 1: TRUE Orange  2
#> 2: TRUE Orange  2
#> 
#> [[37]]
#>     ID bg A1 A2 B1 B2 DR1 DR2 mmA mmB mmDR mmHLA age donor_age dialysis cPRA
#> 1: 487  O  2 33 15 27  11   7   1   2    1     4  55        52      103    0
#> 2: 408  O 29 30 14 57  13   7   2   2    0     4  61        52       93   90
#>       HI     cp SP
#> 1: FALSE Orange  2
#> 2:  TRUE Orange  2
#> 
#> [[38]]
#>     ID bg A1 A2 B1 B2 DR1 DR2 mmA mmB mmDR mmHLA age donor_age dialysis cPRA
#> 1: 317  O  1 68 51 58  13   9   2   1    1     4  31        46      113    0
#> 2: 112  O  2 29 44 44   4  12   1   1    2     4  58        46      104    0
#>       HI     cp SP
#> 1: FALSE Orange  2
#> 2: FALSE Orange  2
#> 
#> [[39]]
#>     ID bg A1 A2 B1 B2 DR1 DR2 mmA mmB mmDR mmHLA age donor_age dialysis cPRA
#> 1: 317  O  1 68 51 58  13   9   1   2    2     5  31        50      113    0
#> 2: 112  O  2 29 44 44   4  12   2   2    2     6  58        50      104    0
#>       HI     cp SP
#> 1: FALSE Orange  2
#> 2: FALSE Orange  2
#> 
#> [[40]]
#>     ID bg A1 A2 B1 B2 DR1 DR2 mmA mmB mmDR mmHLA age donor_age dialysis cPRA
#> 1: 246  A  1 24 15 27  11  13   0   1    1     2  34        30       45   86
#> 2: 326  A 24 32  8 40   3  12   1   2    1     4  41        30       69   86
#>      HI     cp SP
#> 1: TRUE Orange  2
#> 2: TRUE Orange  2
#> 
#> [[41]]
#>     ID bg A1 A2 B1 B2 DR1 DR2 mmA mmB mmDR mmHLA age donor_age dialysis cPRA
#> 1: 246  A  1 24 15 27  11  13   2   2    2     6  34        55       45   86
#> 2:  22  A  1 11 51 57   4   7   1   2    1     4  50        55       71    0
#>       HI     cp SP
#> 1:  TRUE Orange  2
#> 2: FALSE Yellow  2
#> 
#> [[42]]
#>     ID bg A1 A2 B1 B2 DR1 DR2 mmA mmB mmDR mmHLA age donor_age dialysis cPRA
#> 1: 239  A 11 32 35 44   4   7   1   2    1     4  66        68       66    0
#> 2: 442  A  3 25  7 15  13   4   1   2    2     5  68        68       98    0
#>       HI     cp SP
#> 1: FALSE Yellow  1
#> 2: FALSE Yellow  1
#> 
#> [[43]]
#>     ID bg A1 A2 B1 B2 DR1 DR2 mmA mmB mmDR mmHLA age donor_age dialysis cPRA
#> 1: 317  O  1 68 51 58  13   9   1   1    2     4  31        42      113    0
#> 2: 112  O  2 29 44 44   4  12   1   2    2     5  58        42      104    0
#>       HI     cp SP
#> 1: FALSE Orange  2
#> 2: FALSE Orange  2
#> 
#> [[44]]
#>     ID bg A1 A2 B1 B2 DR1 DR2 mmA mmB mmDR mmHLA age donor_age dialysis cPRA
#> 1: 335  O  1 68 35 35  11   8   2   1    0     3  66        68       87    0
#> 2: 130  O  1  2  8 18  11   4   2   1    1     4  72        68       99    0
#>       HI     cp SP
#> 1: FALSE Yellow  1
#> 2: FALSE Yellow  1
#> 
#> [[45]]
#>     ID bg A1 A2 B1 B2 DR1 DR2 mmA mmB mmDR mmHLA age donor_age dialysis cPRA
#> 1: 408  O 29 30 14 57  13   7   2   1    0     3  61        50       93   90
#> 2: 317  O  1 68 51 58  13   9   1   2    1     4  31        50      113    0
#>       HI     cp SP
#> 1:  TRUE Orange  2
#> 2: FALSE Orange  2
#> 
#> [[46]]
#>     ID bg A1 A2 B1 B2 DR1 DR2 mmA mmB mmDR mmHLA age donor_age dialysis cPRA
#> 1: 487  O  2 33 15 27  11   7   0   2    1     3  55        22      103    0
#> 2: 236  O  2 33 14 44  10  10   0   1    2     3  38        22       90   88
#>       HI     cp SP
#> 1: FALSE Orange  2
#> 2:  TRUE Orange  2
#> 
#> [[47]]
#>    ID bg A1 A2 B1 B2 DR1 DR2 mmA mmB mmDR mmHLA age donor_age dialysis cPRA
#> 1: 73 AB  2  2 44 58  11  13   2   2    1     5  71        68       15    0
#>       HI    cp SP
#> 1: FALSE Green  1
#> 
#> [[48]]
#>     ID bg A1 A2 B1 B2 DR1 DR2 mmA mmB mmDR mmHLA age donor_age dialysis cPRA
#> 1: 397  O  1  2 39 57   7   8   1   1    2     4  68        66       83    0
#> 2: 130  O  1  2  8 18  11   4   1   2    2     5  72        66       99    0
#>       HI     cp SP
#> 1: FALSE Yellow  1
#> 2: FALSE Yellow  1
#> 
#> [[49]]
#>     ID bg A1 A2 B1 B2 DR1 DR2 mmA mmB mmDR mmHLA age donor_age dialysis cPRA
#> 1: 246  A  1 24 15 27  11  13   2   2    2     6  34        58       45   86
#> 2: 212  A 23 32 14 44   7   7   1   1    1     3  53        58       84    0
#>       HI     cp SP
#> 1:  TRUE Orange  2
#> 2: FALSE Yellow  2
#> 
#> [[50]]
#>     ID bg A1 A2 B1 B2 DR1 DR2 mmA mmB mmDR mmHLA age donor_age dialysis cPRA
#> 1: 344  A  1  3 35 49   1   4   2   1    0     3  70        67       69    0
#> 2: 215  A  2 25 50 51  11   7   1   2    1     4  68        67       66    0
#>       HI     cp SP
#> 1: FALSE Yellow  1
#> 2: FALSE Yellow  1
#> 
#> [[51]]
#>     ID bg A1 A2 B1 B2 DR1 DR2 mmA mmB mmDR mmHLA age donor_age dialysis cPRA
#> 1: 385  A  2 24 15 35   3  11   1   1    2     4  49        52      103   90
#> 2: 114  A  2 25  7 15  13  13   1   1    1     3  51        52       65   60
#>       HI     cp SP
#> 1:  TRUE Orange  2
#> 2: FALSE Yellow  2
#> 
#> [[52]]
#>     ID bg A1 A2 B1 B2 DR1 DR2 mmA mmB mmDR mmHLA age donor_age dialysis cPRA
#> 1: 464  A  1  3 35 57  13   7   2   2    2     6  52        43       48   88
#> 2:  17  A  1 31  8 18   3  14   2   0    1     3  47        43       88    0
#>       HI     cp SP
#> 1:  TRUE Orange  2
#> 2: FALSE Yellow  2
#> 
#> [[53]]
#>     ID bg A1 A2 B1 B2 DR1 DR2 mmA mmB mmDR mmHLA age donor_age dialysis cPRA
#> 1:  17  A  1 31  8 18   3  14   2   1    1     4  47        62       88    0
#> 2: 243  A  2 24  7 49   4   4   0   2    2     4  22        62       73    0
#>       HI     cp SP
#> 1: FALSE Yellow  2
#> 2: FALSE Yellow  2
#> 
#> [[54]]
#>     ID bg A1 A2 B1 B2 DR1 DR2 mmA mmB mmDR mmHLA age donor_age dialysis cPRA
#> 1: 385  A  2 24 15 35   3  11   1   1    1     3  49        57      103   90
#> 2: 464  A  1  3 35 57  13   7   1   2    0     3  52        57       48   88
#>      HI     cp SP
#> 1: TRUE Orange  2
#> 2: TRUE Orange  2
#> 
#> [[55]]
#>     ID bg A1 A2 B1 B2 DR1 DR2 mmA mmB mmDR mmHLA age donor_age dialysis cPRA
#> 1: 317  O  1 68 51 58  13   9   1   2    2     5  31        63      113    0
#> 2: 487  O  2 33 15 27  11   7   2   2    1     5  55        63      103    0
#>       HI     cp SP
#> 1: FALSE Orange  2
#> 2: FALSE Orange  2
#> 
#> [[56]]
#>     ID bg A1 A2 B1 B2 DR1 DR2 mmA mmB mmDR mmHLA age donor_age dialysis cPRA
#> 1: 112  O  2 29 44 44   4  12   2   2    1     5  58        53      104    0
#> 2: 487  O  2 33 15 27  11   7   2   2    1     5  55        53      103    0
#>       HI     cp SP
#> 1: FALSE Orange  2
#> 2: FALSE Orange  2
#> 
#> [[57]]
#>     ID bg A1 A2 B1 B2 DR1 DR2 mmA mmB mmDR mmHLA age donor_age dialysis cPRA
#> 1: 143  A  2 24  7 44  15   4   1   1    1     3  20        62       73   86
#> 2: 314  A  2  2 44 50  13   4   1   1    0     2  46        62       62    0
#>       HI     cp SP
#> 1:  TRUE Orange  2
#> 2: FALSE Yellow  2
#> 
#> [[58]]
#>     ID bg A1 A2 B1 B2 DR1 DR2 mmA mmB mmDR mmHLA age donor_age dialysis cPRA
#> 1: 143  A  2 24  7 44  15   4   0   1    1     2  20        56       73   86
#> 2: 246  A  1 24 15 27  11  13   0   2    1     3  34        56       45   86
#>      HI     cp SP
#> 1: TRUE Orange  2
#> 2: TRUE Orange  2
#> 
#> [[59]]
#>     ID bg A1 A2 B1 B2 DR1 DR2 mmA mmB mmDR mmHLA age donor_age dialysis cPRA
#> 1: 288  B  3 24  7  7  15  15   1   1    2     4  61        42       22   82
#> 2:   6  B 23 33 14 44   7   7   2   2    1     5  51        42       76    0
#>       HI     cp SP
#> 1: FALSE Yellow  2
#> 2: FALSE Yellow  2
#> 
#> [[60]]
#>     ID bg A1 A2 B1 B2 DR1 DR2 mmA mmB mmDR mmHLA age donor_age dialysis cPRA
#> 1:  28  B  2 68  8 39   3  11   1   1    1     3  34        53       58   62
#> 2: 206  B 26 33 14 38   1  13   1   2    1     4  21        53       60    0
#>       HI     cp SP
#> 1: FALSE Yellow  2
#> 2: FALSE Yellow  2
#> 
#> [[61]]
#>     ID bg A1 A2 B1 B2 DR1 DR2 mmA mmB mmDR mmHLA age donor_age dialysis cPRA
#> 1: 112  O  2 29 44 44   4  12   1   1    1     3  58        26      104    0
#> 2: 317  O  1 68 51 58  13   9   1   2    2     5  31        26      113    0
#>       HI     cp SP
#> 1: FALSE Orange  2
#> 2: FALSE Orange  2
#> 
#> [[62]]
#>     ID bg A1 A2 B1 B2 DR1 DR2 mmA mmB mmDR mmHLA age donor_age dialysis cPRA
#> 1: 314  A  2  2 44 50  13   4   2   0    0     2  46        63       62    0
#> 2: 286  A  3 29 35 44   1   4   1   1    1     3  25        63       72    0
#>       HI     cp SP
#> 1: FALSE Yellow  2
#> 2: FALSE Yellow  2
#> 
#> [[63]]
#>     ID bg A1 A2 B1 B2 DR1 DR2 mmA mmB mmDR mmHLA age donor_age dialysis cPRA
#> 1: 317  O  1 68 51 58  13   9   1   1    2     4  31        43      113    0
#> 2: 112  O  2 29 44 44   4  12   2   2    1     5  58        43      104    0
#>       HI     cp SP
#> 1: FALSE Orange  2
#> 2: FALSE Orange  2
#> 
#> [[64]]
#>     ID bg A1 A2 B1 B2 DR1 DR2 mmA mmB mmDR mmHLA age donor_age dialysis cPRA
#> 1: 317  O  1 68 51 58  13   9   1   2    1     4  31        39      113    0
#> 2: 112  O  2 29 44 44   4  12   1   2    2     5  58        39      104    0
#>       HI     cp SP
#> 1: FALSE Orange  2
#> 2: FALSE Orange  2
#> 
#> [[65]]
#>     ID bg A1 A2 B1 B2 DR1 DR2 mmA mmB mmDR mmHLA age donor_age dialysis cPRA
#> 1: 444  B  3 30 18 35  15  13   1   1    0     2  51        46       35   78
#> 2: 336  B  2  2  7 14   1  13   2   1    1     4  45        46       62   70
#>       HI     cp SP
#> 1: FALSE Yellow  2
#> 2: FALSE Yellow  2
#> 
#> [[66]]
#>     ID bg A1 A2 B1 B2 DR1 DR2 mmA mmB mmDR mmHLA age donor_age dialysis cPRA
#> 1:  16  O  3 24  8 18   3  13   1   1    1     3  66        66       60    0
#> 2: 130  O  1  2  8 18  11   4   2   1    1     4  72        66       99    0
#>       HI     cp SP
#> 1: FALSE Yellow  1
#> 2: FALSE Yellow  1
#> 
#> [[67]]
#>     ID bg A1 A2 B1 B2 DR1 DR2 mmA mmB mmDR mmHLA age donor_age dialysis cPRA
#> 1: 408  O 29 30 14 57  13   7   2   1    0     3  61        62       93   90
#> 2: 487  O  2 33 15 27  11   7   1   2    1     4  55        62      103    0
#>       HI     cp SP
#> 1:  TRUE Orange  2
#> 2: FALSE Orange  2
#> 
#> [[68]]
#>     ID bg A1 A2 B1 B2 DR1 DR2 mmA mmB mmDR mmHLA age donor_age dialysis cPRA
#> 1: 487  O  2 33 15 27  11   7   1   1    1     3  55        53      103    0
#> 2: 112  O  2 29 44 44   4  12   1   2    1     4  58        53      104    0
#>       HI     cp SP
#> 1: FALSE Orange  2
#> 2: FALSE Orange  2
#> 
#> [[69]]
#>     ID bg A1 A2 B1 B2 DR1 DR2 mmA mmB mmDR mmHLA age donor_age dialysis cPRA
#> 1: 487  O  2 33 15 27  11   7   1   1    1     3  55        44      103    0
#> 2: 317  O  1 68 51 58  13   9   2   2    1     5  31        44      113    0
#>       HI     cp SP
#> 1: FALSE Orange  2
#> 2: FALSE Orange  2
#> 
#> [[70]]
#>     ID bg A1 A2 B1 B2 DR1 DR2 mmA mmB mmDR mmHLA age donor_age dialysis cPRA
#> 1: 314  A  2  2 44 50  13   4   1   1    0     2  46        58       62    0
#> 2:  81  A  2 23 15 49  15   4   1   1    1     3  44        58       91    0
#>       HI     cp SP
#> 1: FALSE Yellow  2
#> 2: FALSE Yellow  2
#>