resume function for ET algorithm punctuation

Ordering of waitlisted candidates for a given donor and according to ETKAS algorithm.

Usage

et(
  iso = TRUE,
  dABO = "A",
  dA = c("1", "2"),
  dB = c("15", "44"),
  dDR = c("1", "4"),
  donor.age = 65,
  data = candidates,
  month = 2.78,
  mm0 = 400,
  mm1 = 333.33,
  mm2 = 266.67,
  mm3 = 200,
  mm4 = 133.33,
  mm5 = 66.67,
  mm6 = 0,
  df.abs = cabs,
  hlaA = hlaApt,
  hlaB = hlaBpt,
  hlaDR = hlaDRpt,
  abo_freq = ABOpt,
  n = 2,
  check.validity = TRUE
)

Arguments

iso

A logical value for isogroupal compatibility.

dABO

A character value with ABO blood group (env$valid.blood.groups).

dA

donor's HLA-A typing.

dB

donor's HLA-B typing.

dDR

donor's HLA-DR typing.

donor.age

A numeric value with donor's age.

data

A data frame containing demographics and medical information for a group of waitlisted transplant candidates.

month

A numeric value with the punctuation for each month (between env$month.points.minimum and env$month.points.maximum)

mm0

A numeric value with points for 0 HLA mm on ETKAS points table

mm1

A numeric value with points for 1 HLA mm on ETKAS points table

mm2

A numeric value with points for 2 HLA mm on ETKAS points table

mm3

A numeric value with points for 3 HLA mm on ETKAS points table

mm4

A numeric value with points for 4 HLA mm on ETKAS points table

mm5

A numeric value with points for 5 HLA mm on ETKAS points table

mm6

A numeric value with points for 6 HLA mm on ETKAS points table

df.abs

A data frame with candidates' antibodies.

hlaA

A data frame with HLA-A allele frequencies

hlaB

A data frame with HLA-B allele frequencies

hlaDR

A data frame with HLA-DR allele frequencies

abo_freq

A data frame with ABO blood group frequencies

n

A positive integer to slice the first candidates.

check.validity

Logical to decide whether to validate input.

Value

An ordered data frame with columns cPRA, HI, pointsET, SP, AM and 'mmHLA'.

Examples

et(iso = TRUE, dABO = "A",
dA = c("1","2"), dB = c("15","44"), dDR = c("1","4"),
donor.age = 65,
data = candidates, month = 2.78,
mm0 = 400, mm1 = 333.33, mm2 = 266.67, mm3 = 200,
mm4 = 133.33, mm5 = 66.67, mm6 = 0,
df.abs = cabs,
hlaA = hlaApt, hlaB = hlaBpt, hlaDR = hlaDRpt,
abo_freq = ABOpt,
n = 2,
check.validity = TRUE)
#>     ID bg A1 A2 B1 B2 DR1 DR2 mmA mmB mmDR mmHLA age donor_age dialysis cPRA
#> 1: 442  A  3 25  7 15  13   4   2   1    1     4  68        65       98    0
#> 2: 344  A  1  3 35 49   1   4   1   2    0     3  70        65       69    0
#>       HI pointsET SP AM
#> 1: FALSE       98  1  0
#> 2: FALSE       69  1  0