Mismatch Probability (MMP) from ETKAS

Mismatch Probability (MMP) is a calculation of the probability of receiving a kidney offer with 0 and 1 broad HLA-A, -B or split DR mismatches based on 1000 kidneys offered, taking into account AB0 blood group rules and PRA screening. Patients receive between 0-100 MMPs

Usage

et_mmp(
  data = candidates,
  hlaA = hlaApt,
  hlaB = hlaBpt,
  hlaDR = hlaDRpt,
  abo_freq = ABOpt,
  check.validity = TRUE
)

Source

https://www.eurotransplant.org/wp-content/uploads/2020/01/H4-Kidney.pdf

Arguments

data

A data frame containing demographics and medical information for a group of waitlisted transplant candidates.

hlaA

A data frame with HLA-A allele frequencies

hlaB

A data frame with HLA-B allele frequencies

hlaDR

A data frame with HLA-DR allele frequencies

abo_freq

A data frame with ABO blood group frequencies

check.validity

Logical to decide whether to validate input.

Examples

et_mmp(data = candidates,
hlaA = hlaApt, hlaB = hlaBpt, hlaDR = hlaDRpt,
abo_freq = ABOpt, check.validity = TRUE)
#>      bg    abo DR2   DR2_freq DR1   DR1_freq B2    B2_freq B1    B1_freq A2
#>   1:  A 0.4660   4 0.12983971   1 0.12189087 44 0.15036980 15 0.05647093 24
#>   2:  O 0.4228   7 0.17697997   4 0.12983971 58 0.01871397 57 0.02859737  2
#>   3:  O 0.4228   7 0.17697997   4 0.12983971 44 0.15036980 18 0.04769300 29
#>   4:  O 0.4228  12 0.01538441   1 0.12189087 44 0.15036980 14 0.07571132 31
#>   5:  O 0.4228  15 0.08401548   1 0.12189087 35 0.11773221  7 0.06018214 24
#>  ---                                                                       
#> 496:  O 0.4228  14 0.02570210  13 0.15094886 51 0.10688811 49 0.03394046 68
#> 497:  B 0.0769   8 0.04441608   3 0.10208460 51 0.10688811  8 0.06707815 29
#> 498:  A 0.4660  13 0.15094886   1 0.12189087 27 0.03075567 14 0.07571132 33
#> 499:  O 0.4228   7 0.17697997  16 0.02441239 44 0.15036980 39 0.01734530 33
#> 500:  O 0.4228   4 0.12983971  11 0.10053168 37 0.01308136 35 0.11773221 32
#>         A2_freq A1    A1_freq  ID age dialysis cPRA urgent         MMP0
#>   1: 0.10283473  2 0.27503751   1  53       59    0      0 3.871097e-04
#>   2: 0.27503751  1 0.10894112   2  41       35   16      0 3.106785e-05
#>   3: 0.05353618 25 0.01605559   3  72       53    0      0 1.788496e-05
#>   4: 0.02518885  2 0.27503751   4  48       50    0      0 8.681824e-05
#>   5: 0.10283473  3 0.10362435   5  62       39    0      0 5.720440e-05
#>  ---                                                                   
#> 496: 0.04998289  1 0.10894112 496  48       21    0      0 1.563121e-05
#> 497: 0.05353618  1 0.10894112 497  25       22    0      0 1.714726e-05
#> 498: 0.03511173  2 0.27503751 498  62       62    0      0 8.116832e-05
#> 499: 0.03511173 23 0.04510041 499  58       49    0      0 7.340246e-06
#> 500: 0.03917827  1 0.10894112 500  36       21    0      0 1.992446e-05
#>          MMP2         MMP1       MMP
#>   1: 19.68507 0.0076202820  2.379158
#>   2: 81.97308 0.0025467271 40.014498
#>   3: 66.75502 0.0011939109 59.900904
#>   4: 30.96131 0.0026880059 30.916370
#>   5: 31.16391 0.0017827127 45.922285
#>  ---                                
#> 496: 42.68281 0.0006671839 74.921003
#> 497: 42.51094 0.0007289461 94.423876
#> 498: 34.17044 0.0027735569 26.416335
#> 499: 63.83284 0.0004685488 81.772751
#> 500: 41.87166 0.0008342702 69.682657