resume function for UK algorithm punctuation

Ordering of waitlisted candidates for a given donor and according to UK transplant algorithm.

Usage

uk(
  DRI = "D1",
  dA = c("1", "2"),
  dB = c("15", "44"),
  dDR = c("1", "4"),
  dABO = "O",
  donor.age = 65,
  data = candidates.uk,
  D1R1 = 1000,
  D1R2 = 700,
  D1R3 = 350,
  D1R4 = 0,
  D2R1 = 700,
  D2R2 = 1000,
  D2R3 = 500,
  D2R4 = 350,
  D3R1 = 350,
  D3R2 = 500,
  D3R3 = 1000,
  D3R4 = 700,
  D4R1 = 0,
  D4R2 = 350,
  D4R3 = 700,
  D4R4 = 1000,
  ptsDial = 1,
  a1 = 2300,
  a2 = 1500,
  b1 = 1200,
  b2 = 750,
  b3 = 400,
  m = 40,
  nn = 4.5,
  o = 4.7,
  mm1 = -100,
  mm23 = -150,
  mm46 = -250,
  pts = -1000,
  df.abs = cabs,
  n = 2,
  check.validity = TRUE
)

Arguments

DRI

Donor RisK Index group (env$valid.dris)

dA

donor's HLA-A typing.

dB

donor's HLA-B typing.

dDR

donor's HLA-DR typing.

dABO

A character value with ABO blood group (env$valid.blood.groups).

donor.age

A numeric value with donor's age.

data

A data frame containing demographics and medical information for a group of waitlisted transplant for UK transplant.

D1R1

A numeric value (env$dirj.minimum - env$dirj.maximum) for the combination of indexes DiRj

D1R2

A numeric value (env$dirj.minimum - env$dirj.maximum) for the combination of indexes DiRj

D1R3

A numeric value (env$dirj.minimum - env$dirj.maximum) for the combination of indexes DiRj

D1R4

A numeric value (env$dirj.minimum - env$dirj.maximum) for the combination of indexes DiRj

D2R1

A numeric value (env$dirj.minimum - env$dirj.maximum) for the combination of indexes DiRj

D2R2

A numeric value (env$dirj.minimum - env$dirj.maximum) for the combination of indexes DiRj

D2R3

A numeric value (env$dirj.minimum - env$dirj.maximum) for the combination of indexes DiRj

D2R4

A numeric value (env$dirj.minimum - env$dirj.maximum) for the combination of indexes DiRj

D3R1

A numeric value (env$dirj.minimum - env$dirj.maximum) for the combination of indexes DiRj

D3R2

A numeric value (env$dirj.minimum - env$dirj.maximum) for the combination of indexes DiRj

D3R3

A numeric value (env$dirj.minimum - env$dirj.maximum) for the combination of indexes DiRj

D3R4

A numeric value (env$dirj.minimum - env$dirj.maximum) for the combination of indexes DiRj

D4R1

A numeric value (env$dirj.minimum - env$dirj.maximum) for the combination of indexes DiRj

D4R2

A numeric value (env$dirj.minimum - env$dirj.maximum) for the combination of indexes DiRj

D4R3

A numeric value (env$dirj.minimum - env$dirj.maximum) for the combination of indexes DiRj

D4R4

A numeric value (env$dirj.minimum - env$dirj.maximum) for the combination of indexes DiRj

ptsDial

A numeric value for the points corresponding to each month on dialysis

a1

A numeric value for HLA match and age combined formula: b1*cos(age / env$adulthood.age)+a1

a2

A numeric value for HLA match and age combined formula: b2*cos(age / env$adulthood.age)+a2

b1

A numeric value for HLA match and age combined formula: b1*cos(age / env$adulthood.age)+a1

b2

A numeric value for HLA match and age combined formula: b2*cos(age / env$adulthood.age)+a2

b3

A numeric value for HLA match and age combined formula: b3*sin(age / 50)

m

A numeric value for matchability formula: m * (1 + (MS / nn) ^ o)

nn

A numeric value for matchability formula: m * (1 + (MS / nn) ^ o)

o

A numeric value for matchability formula: m * (1 + (MS / nn) ^ o)

mm1

A numeric value to penalize 1 mm

mm23

A numeric value to penalize 2-3 mm

mm46

A numeric value to penalize 4-6 mm

pts

A negative value with penalization for B candidates

df.abs

A data frame with candidates' antibodies.

n

A positive integer to slice the first candidates.

check.validity

Logical to decide whether to validate input.

Examples

uk(DRI = 'D1', dA = c("1","2"), dB = c("15","44"), dDR = c("1","4"),
dABO = "O", donor.age = 65, data = candidates.uk,
D1R1 = 1000, D1R2 = 700, D1R3 = 350, D1R4 = 0,
D2R1 = 700, D2R2 = 1000, D2R3 = 500, D2R4 = 350,
D3R1 = 350, D3R2 = 500, D3R3 = 1000, D3R4 = 700,
D4R1 = 0, D4R2 = 350, D4R3 = 700, D4R4 = 1000,
ptsDial = 1,
a1 = 2300,  a2 = 1500, b1 = 1200, b2 = 750, b3 = 400,
m = 40, nn = 4.5, o = 4.7,
mm1 = -100, mm23 = -150, mm46 = -250,
pts = -1000,
df.abs = cabs, n = 2, check.validity = TRUE)
#>     ID bg A1 A2 B1 B2 DR1 DR2 matchability mmA mmB mmDR mmHLA age donor_age
#> 1: 324  O  1 11  8 35   3  13       1745.9   1   2    2     5  69        65
#> 2: 104  O  2  3  8 27   3   3       1745.9   1   2    2     5  45        65
#>    dialysis cPRA Tier pointsUK
#> 1:       81    0    A     9999
#> 2:       66    0    A     9999