Candidates' Color Priority

Classification of candidates according to waiting list time on dialysis' quartiles and two cPRA cutoff values.

Usage

cp(data = candidates, q2 = 60, q3 = 100, cPRA1 = 50, cPRA2 = 85)

Arguments

data

A data frame with information for candidates' waiting list.

q2

A numerical value for the median of candidates' waiting list (0 - 120).

q3

A numerical value for the third quartile of candidates' waiting list (0 - 120).

cPRA1

A numerical value (100 - 100) for the lower cPRA cutoff.

cPRA2

A numerical value (100 - 100) for the higher cPRA cutoff. cPRA2 must be greater than cPRA1.

Value

A data frame with a new column 'cp' (color priority)

Examples

cp(data = candidates, q2 = 60, q3 = 100, cPRA1 = 50, cPRA2 = 85)
#>      ID bg A1 A2 B1 B2 DR1 DR2 age dialysis cPRA urgent     cp
#> 1     1  A  2 24 15 44   1   4  53       59    0      0  Green
#> 2     2  O  1  2 57 58   4   7  41       35   16      0  Green
#> 3     3  O 25 29 18 44   4   7  72       53    0      0  Green
#> 4     4  O  2 31 14 44   1  12  48       50    0      0  Green
#> 5     5  O  3 24  7 35   1  15  62       39    0      0  Green
#> 6     6  B 23 33 14 44   7   7  51       76    0      0 Yellow
#> 7     7  O  2 32 14 57   1   3  63       70    0      0 Yellow
#> 8     8  A 11 30 13 35   1  13  49        8    0      0  Green
#> 9     9  B  2 24  7 51  15   4  40       34    0      0  Green
#> 10   10  A  2 24  7  7  15   8  62       33    0      0  Green
#> 11   11  O  1 29  8 44   7   7  29       52   46      0  Green
#> 12   12  A  2  3 18 35  11   4  52       47    0      0  Green
#> 13   13  B  2 11 35 44   7  12  25       13    0      0  Green
#> 14   14  O  2 32 14 51   1  12  38       74    0      0 Yellow
#> 15   15  O  3 25 15 35   1  13  42       89    0      0 Yellow
#> 16   16  O  3 24  8 18   3  13  66       60    0      0 Yellow
#> 17   17  A  1 31  8 18   3  14  47       88    0      0 Yellow
#> 18   18  O  1  1 35 35   4   4  43       62    0      0 Yellow
#> 19   19  O 24 33 18 35   4   7  56       79   48      0 Yellow
#> 20   20  O 24 31  7 51  15   8  49       68    0      0 Yellow
#> 21   21  A  3 11 35 57   4   7  73       41    0      0  Green
#> 22   22  A  1 11 51 57   4   7  50       71    0      0 Yellow
#> 23   23  O  2 30  8 13   3  12  29       81    0      0 Yellow
#> 24   24  O  3 26 14 35  16   3  54       80   82      0 Yellow
#> 25   25  O  2 11 49 51   3   8  47       59    0      0  Green
#> 26   26  O  1 29  8 49   3   7  23       87    0      0 Yellow
#> 27   27  A 11 29 35 44   1  11  29        2   52      0 Yellow
#> 28   28  B  2 68  8 39   3  11  34       58   62      0 Yellow
#> 29   29  O  2 23 18 50   3  11  45       54    0      0  Green
#> 30   30  A  2 24 35 44  13   7  58       14    0      0  Green
#> 31   31  O  2  3  8 44   1   1  39       72    0      0 Yellow
#> 32   32  A  2 11 35 51   1  11  37       41    0      0  Green
#> 33   33  O 24 31 44 51  11   4  41       58    0      0  Green
#> 34   34  O  2 24 15 40   3  13  36       68   52      0 Yellow
#> 35   35  O  2 23 44 57  15   7  40       68    0      0 Yellow
#> 36   36  A  3 24 14 35  11   7  41       24    0      0  Green
#> 37   37 AB  1 24 14 35  13   7  32       64   60      0 Yellow
#> 38   38  B  1 26  7 51  15   4  32       22    0      0  Green
#> 39   39  O  1  3 15 57  13   7  47       69    0      0 Yellow
#> 40   40  O  3 24  7 35  15  15  69       41    0      0  Green
#> 41   41  A  2  2 13 49   4   7  53       26    0      0  Green
#> 42   42  O  2  2  8 39   3   8  69       41    0      0  Green
#> 43   43  A 23 33 14 35  13   7  38       67    0      0 Yellow
#> 44   44  A 24 30 18 37   3  13  34       49    0      0  Green
#> 45   45  O  2  2 41 57  13   4  30       88    0      0 Yellow
#> 46   46  A  1 66 41 51  13   4  49       14    0      0  Green
#> 47   47  A  1  2 44 44  11   4  66       62    0      0 Yellow
#> 48   48  A  2 29 51 53   3  13  49       25    0      0  Green
#> 49   49  O 11 26 38 51  11   7  56       43    0      0  Green
#> 50   50  O  2 11 14 27   7   8  40       67   82      0 Yellow
#> 51   51  A  2 26 38 53  13  13  37       26    0      0  Green
#> 52   52  A 24 68  8 40   3  11  49       49    0      0  Green
#> 53   53  O  3 11 35 57  13   7  41       78    0      0 Yellow
#> 54   54  A  2 24 14 45   1  13  32        1    0      0  Green
#> 55   55  O  2 26 44 44  15   4  57       57    0      0  Green
#> 56   56  A 24 26 13 38  16   7  67       19    0      0  Green
#> 57   57  A 11 29 44 44   7   7  55       38    0      0  Green
#> 58   58  O  2 32 14 27  15   7  26       60    0      0 Yellow
#> 59   59  A 23 33 49 58  11   7  49       35    0      0  Green
#> 60   60  B  2  3 18 44  15  13  43        5    0      0  Green
#> 61   61  A  2 32 14 35  11   7  32       22    0      0  Green
#> 62   62  O 23 32 40 51   1   7  60       63    0      0 Yellow
#> 63   63  A  1  2 49 57   1   4  31       44    0      0  Green
#> 64   64  A 11 32 40 40  15   7  36       46    0      0  Green
#> 65   65 AB  2  2 35 51  11   4  62       63    0      0 Yellow
#> 66   66  A  1  2  8 55  16   3  62        6    0      0  Green
#> 67   67  O  2 24  7 50  15   8  41       41    0      0  Green
#> 68   68  O  2  3  7  7  15  15  29       31    0      0  Green
#> 69   69  O 26 68 49 55  16   4  44       36    0      0  Green
#> 70   70  A  2 68 39 78  16  13  37       61    0      0 Yellow
#> 71   71  A 11 33 14 35   1   1  47       53    0      0  Green
#> 72   72  O 32 68 15 53  16  13  27       37    0      0  Green
#> 73   73 AB  2  2 44 58  11  13  71       15    0      0  Green
#> 74   74 AB  2 11 14 57  11   7  39       52    0      0  Green
#> 75   75  A 31 31 14 57   4   7  52       56    0      0  Green
#> 76   76  O 23 68 44 50   3   7  52        0    0      0  Green
#> 77   77  O  2  3  7 57  16   7  44       27    0      0  Green
#> 78   78 AB  3 29  7 51  13   4  49       54   16      0  Green
#> 79   79  A  2 11 50 53  13  14  27       46   84      0 Yellow
#> 80   80  O  3 29 44 49   3   4  66       39    0      0  Green
#> 81   81  A  2 23 15 49  15   4  44       91    0      0 Yellow
#> 82   82  A 11 24  7 39  15   4  51       49    0      0  Green
#> 83   83  O  2 69 15 44   7  13  29       38    0      0  Green
#> 84   84  A 11 68 38 55   1   4  69       60    0      0 Yellow
#> 85   85  A  1  2 44 57   4   7  60       37    0      0  Green
#> 86   86  O  2 34  8 50   3   7  51       56    0      0  Green
#> 87   87  O 24 68  8 55   4   7  37       43   58      0 Yellow
#> 88   88  A  1  2 51 57   4   8  27       50    0      0  Green
#> 89   89  A  3 24  7 35  15  11  25       57    0      0  Green
#> 90   90  A  3 68 18 53   1  13  24       14    0      0  Green
#> 91   91  A  2 68 51 51  13  13  41       34    0      0  Green
#> 92   92  A 25 25 15 15   1   7  27       29    0      0  Green
#> 93   93  A 11 32 35 40   1  13  48       22   72      0 Yellow
#> 94   94  O 31 32 40 51   1  13  46       92   90      0 Orange
#> 95   95  O  2 32 14 44  14  14  58       60    0      0 Yellow
#> 96   96  O  3 26 18 38  11   8  39       48    0      0  Green
#> 97   97  O  3 23 49 49   4   8  54       51    0      0  Green
#> 98   98  O  3 29 35 44  11   7  21       79    0      0 Yellow
#> 99   99  O 26 31 51 51  13   4  47       93    0      0 Yellow
#> 100 100  O  3 30 18 35   3   7  41       56   78      0 Yellow
#> 101 101  A  2 24 35 58  13   4  37       17    0      0  Green
#> 102 102  O  2  3 18 58  15  15  63       90    0      0 Yellow
#> 103 103  O  2 11 44 51  13   4  51       50    0      0  Green
#> 104 104  O  2  3  8 27   3   3  45       66    0      0 Yellow
#> 105 105  A  1  2  8  8   3   4  70       50    0      0  Green
#> 106 106  A 24 25  7 18  15  15  32       29    0      0  Green
#> 107 107  A  3 24 14 40   1   7  70       38    0      0  Green
#> 108 108  A  2 23 39 58   7   8  30        0    0      0  Green
#> 109 109  A  1  1  8 44   1   3  33       22    0      0  Green
#> 110 110  O 23 33 14 35   1  14  44       35    0      0  Green
#> 111 111  O 24 29 35 57  11   7  43       61   70      0 Yellow
#> 112 112  O  2 29 44 44   4  12  58      104    0      0 Orange
#> 113 113  O  3 11  7 27  15   4  56       66   26      0 Yellow
#> 114 114  A  2 25  7 15  13  13  51       65   60      0 Yellow
#> 115 115  A  3 25  7 49  15  13  42       42    0      0  Green
#> 116 116  A 11 30 14 18   3   7  65       49    0      0  Green
#> 117 117  O  3 33 14 15   1  13  52       54    0      0  Green
#> 118 118  A  1 25 15 55  11  13  44       23    0      0  Green
#> 119 119  O  2  3 49 51  13   8  44       38    0      0  Green
#> 120 120  O  1 11 15 45   1  15  25       80    0      0 Yellow
#> 121 121  A 11 26 35 55  15   7  42       45    0      0  Green
#> 122 122  A  1  2  8 40  11  13  41       11    0      0  Green
#> 123 123  A  2 11 35 40  11  10  44       41    0      0  Green
#> 124 124  A  2 24 41 52  15   3  30       27   88      0 Orange
#> 125 125  A  2 68 18 37  11   8  32       27    0      0  Green
#> 126 126  B  1  2 35 44   4   4  58       48   86      0 Orange
#> 127 127  O  1 24 45 57   3  10  47       41   82      0 Yellow
#> 128 128  A  2 30 27 44   7   8  51       51   38      0  Green
#> 129 129  A  2 24 40 45  14  10  59       75   90      0 Orange
#> 130 130  O  1  2  8 18  11   4  72       99    0      0 Yellow
#> 131 131  A  1 33  8 14   1   3  42       56   82      0 Yellow
#> 132 132  A 23 33 14 44   3   7  39       16    0      0  Green
#> 133 133  O 23 29 44 44   4   7  56       39    0      0  Green
#> 134 134  O  3  3 49 51  15   4  59       62    0      0 Yellow
#> 135 135  A 11 11  7 35   1   4  26       53    0      0  Green
#> 136 136  B 24 24 15 39   1   8  50       37   88      0 Orange
#> 137 137  O 24 68 27 35   1  11  28       59   24      0  Green
#> 138 138  O  1  2  8 18  11  13  29       58    0      0  Green
#> 139 139  A  1  2  8 40   3   8  51       31    0      0  Green
#> 140 140  O  1  2 37 41  13   4  71       48    0      0  Green
#> 141 141  O  2 29 38 44   4   8  21       93   86      0 Orange
#> 142 142  A  1 11 37 51  13   8  40       44    0      0  Green
#> 143 143  A  2 24  7 44  15   4  20       73   86      0 Orange
#> 144 144  O 11 31 14 35   1   1  28       73    0      0 Yellow
#> 145 145  B 24 32 15 39  13   8  71       44   86      0 Orange
#> 146 146 AB  2 30 14 44   1  12  27       30    0      0  Green
#> 147 147  B  2 68 18 51   3   4  40       36    0      0  Green
#> 148 148  O  2 69 35 50   1  11  44       66    0      0 Yellow
#> 149 149  A 11 33 14 51   1  12  32       30    0      0  Green
#> 150 150  A  1  2 35 51  11  11  49       15    0      0  Green
#> 151 151  O  3  3 18 35  13  13  45       19    0      0  Green
#> 152 152  O  1 25  8 15   3  13  46       24    0      0  Green
#> 153 153  O 33 33 14 35   3  14  62        0    0      0  Green
#> 154 154  A  2 24 27 58   1   7  39       32    0      0  Green
#> 155 155  O  2 66 40 57  15   4  55       58    0      0  Green
#> 156 156  A  2  3  7 58  11   4  54       46    0      0  Green
#> 157 157  O 32 33 14 14   1  15  30       48    0      0  Green
#> 158 158  O  1  2  8 40   3  11  49        6    0      0  Green
#> 159 159  O 30 68 15 40  13  13  45       62    0      0 Yellow
#> 160 160  B  2 11 44 49   7   7  57       48    0      0  Green
#> 161 161  O  2 24 35 51  11   8  25       42    0      0  Green
#> 162 162  O  1 68  8 51   3   7  40       91    0      0 Yellow
#> 163 163  O  2 24 35 38  11  13  60       18    0      0  Green
#> 164 164 AB 11 26 38 51   7  10  57       24    0      0  Green
#> 165 165  B 24 31 35 44   1  12  50       41    0      0  Green
#> 166 166  O  1  3  8 15  15   3  31       68    0      0 Yellow
#> 167 167  A  1 31  8 40  11  14  57       50    0      0  Green
#> 168 168  A  3 29 35 44  15  13  38       31    0      0  Green
#> 169 169  A 24 68  8 44   3   7  37       18    0      0  Green
#> 170 170  O  3 68 49 51  11  13  55       79    0      0 Yellow
#> 171 171  O  2 32 27 51   4   4  42       51    0      0  Green
#> 172 172  O  3 24  7 35  11  14  48       29    0      0  Green
#> 173 173  A  2 24 35 40   1   3  26       24    0      0  Green
#> 174 174  A  1 24 35 39   1  13  48       75    0      0 Yellow
#> 175 175  O 11 24  7 35  15  11  41       64    0      0 Yellow
#> 176 176  A  1 24  7 35  15   4  29       50    0      0  Green
#> 177 177  A  3 68 35 38  15   7  57       31    0      0  Green
#> 178 178  O 11 29 15 56   1  10  60       73    0      0 Yellow
#> 179 179  A  3 11 35 45  11  10  52       35    0      0  Green
#> 180 180  O  2 68 35 51  11   4  52       58    0      0  Green
#> 181 181  B 24 68 50 51  11   7  29        3    0      0  Green
#> 182 182  O 11 24 35 40   1   7  54       47    0      0  Green
#> 183 183  A  1  3 14 40   1  15  45       56   82      0 Yellow
#> 184 184  O  2 30 27 58  13   8  56       68    0      0 Yellow
#> 185 185  O  2 11 27 44   1  13  64       21    0      0  Green
#> 186 186  A  2 24 15 15  13   8  34       45    0      0  Green
#> 187 187  O  2 33  7 14   1  13  44       55    0      0  Green
#> 188 188  O  2 24 44 51   7  12  60       49    0      0  Green
#> 189 189  A  2 24 51 53  11  13  57       55    0      0  Green
#> 190 190  A 23 29 44 44   1   3  38       42    0      0  Green
#> 191 191  A 24 25 15 18  15   3  62       23    0      0  Green
#> 192 192  O  3 33 14 35   1  16  64       61   94      0 Orange
#> 193 193  A  2 29 15 44   4   7  36       17    0      0  Green
#> 194 194  B  2 32 14 58   1  13  24       46   90      0 Orange
#> 195 195  O  2  2 14 51   3  13  70       67    0      0 Yellow
#> 196 196  A 32 68 27 35  16   4  25       31    0      0  Green
#> 197 197  A  2 11 51 51  11  13  31       44    0      0  Green
#> 198 198 AB  2  3 40 44   1  13  48       27    0      0  Green
#> 199 199  O  3 66 41 44   4   7  32       50   80      0 Yellow
#> 200 200  A  2  3 51 51  11  13  52       58    0      0  Green
#> 201 201  O  2 11 35 51  15   4  20       32    0      0  Green
#> 202 202  O 24 33 15 44   4   7  64       46    0      0  Green
#> 203 203  B  1  1  8 40  16   3  29       21    0      0  Green
#> 204 204  A 29 68 50 53   3  13  49       35    0      0  Green
#> 205 205  A  2  3 35 50   3  11  43       62   96      0 Orange
#> 206 206  B 26 33 14 38   1  13  21       60    0      0 Yellow
#> 207 207  A  1  1  8 37   3  13  36       50    0      0  Green
#> 208 208  O  2 31 37 37   3  13  43       46    0      0  Green
#> 209 209  A  1  2  7 18   7   8  40       46    0      0  Green
#> 210 210  O  2  3  7 37  13   8  56       56    0      0  Green
#> 211 211  A  3 29  7 44  11   7  38       53    0      0  Green
#> 212 212  A 23 32 14 44   7   7  53       84    0      0 Yellow
#> 213 213  O  3 24 44 44   1   7  37       44    0      0  Green
#> 214 214  B  1  3 51 57   4   7  47        0    0      0  Green
#> 215 215  A  2 25 50 51  11   7  68       66    0      0 Yellow
#> 216 216  A  1 23  8 49  15   7  53       35    0      0  Green
#> 217 217  O 11 34  7 55   1  11  45       71   82      0 Yellow
#> 218 218  O  2 24 18 50   7   7  22       91    0      0 Yellow
#> 219 219  O  3 11  7 35   1  16  42       64    0      0 Yellow
#> 220 220  A  3 30 14 35  13  14  56       28   82      0 Yellow
#> 221 221  O 29 31 37 44   7  10  42       78    0      0 Yellow
#> 222 222  O  2 24 35 51   4   7  37       73    0      0 Yellow
#> 223 223  O  2 24  7 50  13   7  27       46    0      0  Green
#> 224 224  O  3 32 35 51   7  12  32       96    0      0 Yellow
#> 225 225  B  2 11 35 51   1  13  37       55    0      0  Green
#> 226 226  O 26 31 14 53   1  13  48       51    0      0  Green
#> 227 227  A  2  3 35 58   1  13  48       42    0      0  Green
#> 228 228  O  2  2 44 44  13  12  28       32    0      0  Green
#> 229 229  O  2 30 13 44  13   8  36       29    0      0  Green
#> 230 230  O  2 69 44 50  14   4  53       80    0      0 Yellow
#> 231 231 AB  2  2 18 51   3  13  39       42    0      0  Green
#> 232 232  O  2 11 35 57   1   7  40       89    0      0 Yellow
#> 233 233  A  2 31 35 51  11  13  42       53    0      0  Green
#> 234 234  O  3 29  7 51  11   4  54       65    0      0 Yellow
#> 235 235  A  1 30  8 57   3  13  51       30   18      0  Green
#> 236 236  O  2 33 14 44  10  10  38       90   88      0 Orange
#> 237 237  A 23 29 45 45  14   7  35       27    0      0  Green
#> 238 238  A  2 32 27 51   7   8  48       36    0      0  Green
#> 239 239  A 11 32 35 44   4   7  66       66    0      0 Yellow
#> 240 240  O  2 23 14 51   1   1  46       62    0      0 Yellow
#> 241 241  A  2 31 49 57   7   9  28       54    0      0  Green
#> 242 242  O  2 11 18 35   1  11  54       81   94      0 Orange
#> 243 243  A  2 24  7 49   4   4  22       73    0      0 Yellow
#> 244 244  A  2 68 35 51   3  11  51       28    0      0  Green
#> 245 245  A 68 69 15 38  13  13  45       33    0      0  Green
#> 246 246  A  1 24 15 27  11  13  34       45   86      0 Orange
#> 247 247  A 23 23 44 50  14   7  55       41    0      0  Green
#> 248 248  B 24 68 44 53  13   4  63       22   40      0  Green
#> 249 249  B  1  2  8 50   3   7  40       48   82      0 Yellow
#> 250 250  A  2 33 14 27  11   8  34       37    0      0  Green
#> 251 251  A  1 24  7 58  15  13  26       47    0      0  Green
#> 252 252  O  2 29 15 49   7  12  46       85    0      0 Yellow
#> 253 253  O  1  2  8 51   3   7  38       54    0      0  Green
#> 254 254  A  2 29  7 58   1   8  19       83   82      0 Yellow
#> 255 255  O  2 68 44 44  13  13  23       64    0      0 Yellow
#> 256 256  A 29 30 44 49  15   4  64       42   30      0  Green
#> 257 257  A  3 24  7 14   1   4  56       48   34      0  Green
#> 258 258  O 29 30 18 44  15   3  31       52    0      0  Green
#> 259 259  O  3 32 27 35   1   7  67      107   90      0 Orange
#> 260 260  O  3 32 27 35   1   1  37       60    0      0 Yellow
#> 261 261 AB  3  3 14 51  15   3  55       28    0      0  Green
#> 262 262  O  2 24  7 44   4   8  43       23   44      0  Green
#> 263 263  O  2 24 18 51  15  15  37       72   80      0 Yellow
#> 264 264  O 11 29 35 39  15  14  46       15   26      0  Green
#> 265 265  O 24 29  7 38  11   4  23       15    0      0  Green
#> 266 266  A  1  2  8 51   3  11  64       54    0      0  Green
#> 267 267  B  3 11 14 15   1  11  51       53    0      0  Green
#> 268 268  A  1 32  8 27   3  11  47       18    0      0  Green
#> 269 269  O  2 11  7 44  13  15  46       57    0      0  Green
#> 270 270  A 11 31 14 51   1   1  49       47    0      0  Green
#> 271 271  O  2 24 35 57  15   7  47       50    0      0  Green
#> 272 272  O  2 23  7 44  13   7  26       11    0      0  Green
#> 273 273  O  2  3 14 57   1   7  38       21    0      0  Green
#> 274 274  O  2 29 35 38   3  14  21       60    0      0 Yellow
#> 275 275  A  1  3  7 50  14   7  26       28   42      0  Green
#> 276 276  A 23 24 39 44   7  10  43       11    0      0  Green
#> 277 277  A  1  2  8 35   7   7  26       19    0      0  Green
#> 278 278  B  2  2 44 51   7  12  28       54    0      0  Green
#> 279 279  O  2 25 35 35  11   7  47       74    0      0 Yellow
#> 280 280  A 23 24 45 50   3   3  50       11    0      0  Green
#> 281 281  O 11 23 14 44   7   7  29       97    0      0 Yellow
#> 282 282  A  1 29  8 35  11  13  48       52    0      0  Green
#> 283 283  A  1 32 44 58  11  11  51       57    0      0  Green
#> 284 284  A  2  2 50 51   7   8  67       34    0      0  Green
#> 285 285  O  3 24  7 27   1  11  28       43    0      0  Green
#> 286 286  A  3 29 35 44   1   4  25       72    0      0 Yellow
#> 287 287  O  2 29 44 44  15   7  61       76    0      0 Yellow
#> 288 288  B  3 24  7  7  15  15  61       22   82      0 Yellow
#> 289 289  O  3  3 44 51  13   4  42       72    0      0 Yellow
#> 290 290 AB  1 68  8 35  16   3  62        4    0      0  Green
#> 291 291  O  2  3 35 44   1  11  69       44    0      0  Green
#> 292 292  O  1 23  8 38   1   3  29       56    0      0  Green
#> 293 293  A 24 32 44 55  13  14  52       49    0      0  Green
#> 294 294  A  1  2 15 57   1   4  50       42    0      0  Green
#> 295 295  A  3 11  7 44  15  13  47       48    0      0  Green
#> 296 296  O  2  3 38 44  13   4  50       46    0      0  Green
#> 297 297  O  2 31 51 51  16   8  55       68    0      0 Yellow
#> 298 298  A  3 33 51 51  15  16  31       73    0      0 Yellow
#> 299 299  O  2 24 35 35   1  14  45       46    0      0  Green
#> 300 300  O 24 26  8 44   3  13  34       49    0      0  Green
#> 301 301  A  1 29 51 51  13   8  43        7    0      0  Green
#> 302 302  O  1  3  8 47   3  13  57       29    0      0  Green
#> 303 303  A  2  2 14 45   1   7  31       46   56      0 Yellow
#> 304 304  O  1 25  8 18   1   3  65       89    0      0 Yellow
#> 305 305  O  2 26 39 51   7   7  53       48    0      0  Green
#> 306 306  O  1  1  8 58   1  13  65       18    0      0  Green
#> 307 307  O  2 11 13 35   7  12  59       81    0      0 Yellow
#> 308 308  B  2 29 44 51  13   8  58        0   34      0  Green
#> 309 309  O  1  2  8 27  13   7  55       55   82      0 Yellow
#> 310 310  O 31 68 51 57  15  13  52       62    0      0 Yellow
#> 311 311  A  1 32 14 18   7   7  68       49   26      0  Green
#> 312 312  A  2  3 41 44  13   4  58        4    0      0  Green
#> 313 313  A  3 11  7  7   4   4  66       59    0      0  Green
#> 314 314  A  2  2 44 50  13   4  46       62    0      0 Yellow
#> 315 315  A  2 26 56 57   3   8  38       15    0      0  Green
#> 316 316  A 24 30  8 18   1  11  30       49    0      0  Green
#> 317 317  O  1 68 51 58  13   9  31      113    0      0 Orange
#> 318 318  A  1  2 18 44  11   7  66       61    0      0 Yellow
#> 319 319  A  2 26 44 51   3  13  47       24    0      0  Green
#> 320 320  O  3 11  7 35  15  14  52      103   80      0 Orange
#> 321 321  O 11 32 51 51  12   9  23       57    0      0  Green
#> 322 322  A  2 26 40 58   7   8  68       13    0      0  Green
#> 323 323  A 24 29 27 44   4   7  47       52    0      0  Green
#> 324 324  O  1 11  8 35   3  13  69       81    0      0 Yellow
#> 325 325  A  1  2  8 50   3  14  46       31    0      0  Green
#> 326 326  A 24 32  8 40   3  12  41       69   86      0 Orange
#> 327 327  O  2 66 15 67   1   7  42       47    0      0  Green
#> 328 328  B  3 33 38 51  11  12  27       70    0      0 Yellow
#> 329 329  A  1  2 38 44  11  11  63        0    0      0  Green
#> 330 330  O 68 68 35 51  13   4  59       76    0      0 Yellow
#> 331 331  O  3 23  7 44   1   1  48       33   18      0  Green
#> 332 332  B  2 32 27 27  13   8  47       17    0      0  Green
#> 333 333  O  2 33 35 44   3  13  51       52    0      0  Green
#> 334 334  A 11 11 55 55  14  14  75       37   12      0  Green
#> 335 335  O  1 68 35 35  11   8  66       87    0      0 Yellow
#> 336 336  B  2  2  7 14   1  13  45       62   70      0 Yellow
#> 337 337  O  1  2 37 58  15  13  47       31    0      0  Green
#> 338 338  A  2 23 18 44  11  13  30       31    0      0  Green
#> 339 339 AB  1  3  8 49   3  13  51       39    0      0  Green
#> 340 340  A  1 11 18 35   1  13  56       63    0      0 Yellow
#> 341 341  B  2  3 35 45   1  13  50       49    0      0  Green
#> 342 342  O  2 29 44 51   7   9  34       36    0      0  Green
#> 343 343  A  3 11 15 27  11   4  42       38    0      0  Green
#> 344 344  A  1  3 35 49   1   4  70       69    0      0 Yellow
#> 345 345  A 26 33 38 57  16  13  67       43    0      0  Green
#> 346 346  A  1 29  7 39  13   7  62       19    0      0  Green
#> 347 347  O 23 24 14 15   1   4  47       41    0      0  Green
#> 348 348  O  1  2 44 51  13   4  55        6   42      0  Green
#> 349 349  O  1  2 49 51  13   4  36       32   50      0 Yellow
#> 350 350  A  1  3  8 51  13  12  47       40   18      0  Green
#> 351 351  A  1  1  8 44   1   7  59       36   88      0 Orange
#> 352 352  O  3 68 35 44  11  11  48       65    0      0 Yellow
#> 353 353  A  2 29  7 51   1  13  43       61    0      0 Yellow
#> 354 354 AB  2  2  7 53  15  13  37       18    0      0  Green
#> 355 355  A  2 29 15 44   3   7  44       39    0      0  Green
#> 356 356  A  3 24 13 14   1  11  26       39    0      0  Green
#> 357 357  O 23 29  8 58   3   7  50       92    0      0 Yellow
#> 358 358  O  2 23 50 50   3  14  70       55    0      0  Green
#> 359 359  O  2  2 15 58  11  13  42       54    0      0  Green
#> 360 360  O 30 31 44 51   7   7  47       51    0      0  Green
#> 361 361  O  2  3  7  7  15  15  50       45    0      0  Green
#> 362 362 AB 26 26 51 51  11   4  59       26    0      0  Green
#> 363 363  A 31 68 35 35   1   7  41       41    0      0  Green
#> 364 364  O  2  2 14 41   3   3  49       45    0      0  Green
#> 365 365  A  2 30 15 39   8   8  69        0    0      0  Green
#> 366 366  A  2 32 35 44   4  12  33       67    0      0 Yellow
#> 367 367  O  2 11 44 51   3   7  62       77    0      0 Yellow
#> 368 368  A  2  2 35 49   4   7  45       26    0      0  Green
#> 369 369  O  2  3 13 15   7   8  34       55    0      0  Green
#> 370 370  O  3 33 14 35   1  13  48       40    0      0  Green
#> 371 371  O 31 68 27 57   4   7  47       62    0      0 Yellow
#> 372 372  O  2  3 14 35   1   1  55       75    0      0 Yellow
#> 373 373  A  2  2 51 57   3   8  31       30    0      0  Green
#> 374 374  O  3 26 40 57   4   7  36       84   82      0 Yellow
#> 375 375  O 23 68 44 49  13   4  45       37   82      0 Yellow
#> 376 376  O  2  3  7 51  15   3  44       63    0      0 Yellow
#> 377 377  A  2 24 50 56   1  12  46       52    0      0  Green
#> 378 378  O  2  2 50 50   7   7  39       34    0      0  Green
#> 379 379  A  2 11 44 51  16  13  51       26    0      0  Green
#> 380 380  A  2 11 35 44  15   7  43       51    0      0  Green
#> 381 381  O  1 24 35 35  16   4  70       31    0      0  Green
#> 382 382  O  3 31  7 51  15   8  72       27    0      0  Green
#> 383 383  O  1 68  8 51  16  13  46       31    0      0  Green
#> 384 384  O 25 32 35 37   4  10  56       31    0      0  Green
#> 385 385  A  2 24 15 35   3  11  49      103   90      0 Orange
#> 386 386  A  1  3 44 51   4   7  47       71    0      0 Yellow
#> 387 387  O  1 11 44 56  13   7  50       17    0      0  Green
#> 388 388  A  2 24 15 44  13   4  53       33   14      0  Green
#> 389 389  A  1 24 15 27   4   4  42       51    0      0  Green
#> 390 390  A  3 68 35 35   1   1  36       32    0      0  Green
#> 391 391  A  2 68 15 51  14   4  19       26   32      0  Green
#> 392 392  O 11 24 18 35   1  11  22       66    0      0 Yellow
#> 393 393  O  2 24 35 50  15  11  63       82   80      0 Yellow
#> 394 394  O  2  2 35 40   3  11  29       39    0      0  Green
#> 395 395  A 23 26  8 44  11   7  23       48    0      0  Green
#> 396 396  A  2  3 18 35  13  13  53        1    0      0  Green
#> 397 397  O  1  2 39 57   7   8  68       83    0      0 Yellow
#> 398 398  A  3 23 40 50   1  14  38       72   84      0 Yellow
#> 399 399  O  1 26 35 50   7   7  55       79   80      0 Yellow
#> 400 400  O  2 26 44 44   4   8  59       31    0      0  Green
#> 401 401  A  1 29  7 57   1   7  57       62    0      0 Yellow
#> 402 402  B 11 30 14 35   1   4  36       60    0      0 Yellow
#> 403 403  O 24 26 44 50  11   4  41       30    0      0  Green
#> 404 404  A  1 23 38 45  11   4  56       55    0      0  Green
#> 405 405  O  1 11 14 35   1   7  48       34    0      0  Green
#> 406 406  O  2  2 15 44   4   7  60       59   82      0 Yellow
#> 407 407  O  3 33 14 35   1   1  55       27    0      0  Green
#> 408 408  O 29 30 14 57  13   7  61       93   90      0 Orange
#> 409 409  O  3 24 37 47   4   4  61       44    0      0  Green
#> 410 410  A  1 24 14 35   1  13  53        6    0      0  Green
#> 411 411  O 24 68 50 53  13   7  65       48    0      0  Green
#> 412 412  A  3 29 35 44  11   7  30       27    0      0  Green
#> 413 413  O  2 29 44 57   3   7  49       68    0      0 Yellow
#> 414 414  A  3 24 15 49  11  13  27       60    0      0 Yellow
#> 415 415  B  3 24 44 51   1  13  39       37    0      0  Green
#> 416 416  O 23 24 35 58  15  13  64       47    0      0  Green
#> 417 417  A 11 68 44 56   1  13  37       71    0      0 Yellow
#> 418 418  O  1  2  8 50   3  13  41       43    0      0  Green
#> 419 419  B  1 31 40 57  16   4  28       35    0      0  Green
#> 420 420  A  2 26 38 49  13   4  36       43    0      0  Green
#> 421 421  A  2 24 14 35   1   1  35       20   24      0  Green
#> 422 422  O 23 32 40 44   1   1  44       29    0      0  Green
#> 423 423  A  2  3 37 53  13  14  43       55    0      0  Green
#> 424 424  O 23 23 44 44   7   7  67       37   94      0 Orange
#> 425 425  O  3 68 35 44   4   7  47       60   14      0 Yellow
#> 426 426  B  1 11  7 15  15   4  30        0    0      0  Green
#> 427 427  A  2  3  8 44   3   7  37       47    0      0  Green
#> 428 428  B  1 33 14 57   1   7  29       50    0      0  Green
#> 429 429  O  3 33 14 35  13   7  70       47    0      0  Green
#> 430 430  O 11 68 14 27   1   1  51       71    0      0 Yellow
#> 431 431  A  1  1 15 15  13   7  27       28   54      0 Yellow
#> 432 432  A  2  3 18 35   1   8  50       23    0      0  Green
#> 433 433  B 29 68 44 44   3   8  55       21   12      0  Green
#> 434 434  O  2  2 51 51  15   4  47       50    0      0  Green
#> 435 435  O  3 24 14 51   1  14  40       82    0      0 Yellow
#> 436 436  A 24 24 49 51  15   4  66       62   96      0 Orange
#> 437 437  O  3 24  7 57  15   4  28       62    0      0 Yellow
#> 438 438  A 31 68 35 45  11   4  54       41    0      0  Green
#> 439 439  A  2 29 53 57   1   7  31       55    0      0  Green
#> 440 440  A  2  2 39 50   7   8  64       40    0      0  Green
#> 441 441  A  2  2 27 49   7   8  54       49    0      0  Green
#> 442 442  A  3 25  7 15  13   4  68       98    0      0 Yellow
#> 443 443 AB 32 68 49 49   1  13  62       27    0      0  Green
#> 444 444  B  3 30 18 35  15  13  51       35   78      0 Yellow
#> 445 445  A  2 66 49 51   4  11  40       81   98      0 Orange
#> 446 446  A  2 24 15 51  13   4  67       64    0      0 Yellow
#> 447 447  O  2 24 15 35  11   7  33       48   50      0 Yellow
#> 448 448  O  2 68 18 50   4   7  54       67    0      0 Yellow
#> 449 449  O  2 66 38 56  13   8  42       80   58      0 Yellow
#> 450 450  O  1 32  8 14   3  11  48       48   46      0  Green
#> 451 451  O  1  2  8 44  15   3  37       64   80      0 Yellow
#> 452 452  B  2  3 35 51  11   7  48       57    0      0  Green
#> 453 453  B  2 66 35 44   4  12  41       31    0      0  Green
#> 454 454  A  2 32 15 40  15  15  39       29    0      0  Green
#> 455 455  A  2 33 14 58   3  10  20        8   42      0  Green
#> 456 456  A 24 24 40 44   1  13  29       37    0      0  Green
#> 457 457  A  3 25  7 18  15  15  39       29    0      0  Green
#> 458 458  O  1  1  8 39   3  13  53       66   80      0 Yellow
#> 459 459  A  3 66 41 49  13   4  28       30   80      0 Yellow
#> 460 460  O  3 30 39 40  13   7  38       69   20      0 Yellow
#> 461 461  A  1  2 44 44  15   7  47       20    0      0  Green
#> 462 462  B  2 11 13 51  16  12  48       23   44      0  Green
#> 463 463  O 11 25 35 40  15   4  29       54    0      0  Green
#> 464 464  A  1  3 35 57  13   7  52       48   88      0 Orange
#> 465 465  A  2  2 35 44   1   3  59       57    0      0  Green
#> 466 466  A 24 26  7 38  13  13  20       30    0      0  Green
#> 467 467  A 26 31 47 55   1  11  44       78   80      0 Yellow
#> 468 468  O  2 32 15 35   1  13  73       99   82      0 Yellow
#> 469 469  A  1 11  8 38   3   8  26       19    0      0  Green
#> 470 470  O  2 29 44 44   4   7  65       77    0      0 Yellow
#> 471 471  O  2 25 35 51   1   4  32       59    0      0  Green
#> 472 472  A 24 25 18 35  15  11  32       25    0      0  Green
#> 473 473  O  1 33 27 44   1   1  25       58    0      0  Green
#> 474 474  A  2  2 27 50  15   4  31       36    0      0  Green
#> 475 475  A  1  2  8 15   3   4  26       23    0      0  Green
#> 476 476  A  1 24  8 15   3  13  40       44    0      0  Green
#> 477 477  A 11 33 14 14   1  11  67        0    0      0  Green
#> 478 478  O  1 26 14 57   7   7  42       54    0      0  Green
#> 479 479  A 23 33 14 44   1   1  48       84   82      0 Yellow
#> 480 480  A 24 24 39 51  15  13  39       48    0      0  Green
#> 481 481  A  1 25  7 18  15  15  35        0    0      0  Green
#> 482 482  A  1  3  7 35   1   4  53       61   38      0 Yellow
#> 483 483  O  1 26 49 58  13   7  66       48    0      0  Green
#> 484 484  A  2 26 35 51   4  12  49       51    0      0  Green
#> 485 485  O  2 31 18 51   3   4  29       45   16      0  Green
#> 486 486  O  3  3 14 35   1   7  45       45    0      0  Green
#> 487 487  O  2 33 15 27  11   7  55      103    0      0 Orange
#> 488 488  A 11 32 35 40  11   4  48       70   14      0 Yellow
#> 489 489  O  2  3  7 14   1  15  34       81   18      0 Yellow
#> 490 490  O  1 29  8 44   4   7  25       45   10      0  Green
#> 491 491  A 23 29 40 44   4   7  47       41    0      0  Green
#> 492 492  A  2 31  7 51  15   4  59       48    0      0  Green
#> 493 493  O  2 32 44 44   4   7  19       74    0      0 Yellow
#> 494 494  O  2 32 35 50  11   7  37       59    0      0  Green
#> 495 495  B  2 26 38 58   1  13  32       47    0      0  Green
#> 496 496  O  1 68 49 51  13  14  48       21    0      0  Green
#> 497 497  B  1 29  8 51   3   8  25       22    0      0  Green
#> 498 498  A  2 33 14 27   1  13  62       62    0      0 Yellow
#> 499 499  O 23 33 39 44  16   7  58       49    0      0  Green
#> 500 500  O  1 32 35 37  11   4  36       21    0      0  Green