Runs several time the function donor_recipient_pairs() for as bootstrap.

Generic function that runs the matchability between all combinations of donors and candidates. Runs an arbitrary number of times ('iteration.number') to provide statistics.

Usage

several(
  iteration.number = 10,
  df.donors = donors,
  df.candidates = candidates,
  df.abs = cabs,
  algorithm = lima,
  n = 0,
  seed.number = 123,
  check.validity = TRUE,
  ...
)

Arguments

iteration.number

Number of times the matchability runs.

df.donors

A data frame containing demographics and medical information for a poll of donors. For uk algorithm must have respective columns.

df.candidates

A data frame containing demographics and medical information for a group of waitlisted transplant candidates. For uk algorithm must have respective columns.

df.abs

A data frame with candidates' antibodies.

algorithm

The name of the function to use. Valid options are: lima, et, pts, uk (without quotation)

n

A positive integer to slice the first candidates.

seed.number

Seed for new random number. seed.number can be NA in which case no seed is applied.

check.validity

Logical to decide whether to validate input.

...

all the parameters used on the function algorithm

Value

Statistics related to all the times the function ran.

Examples

several(iteration.number = 10,
df.donors = donors,
df.candidates = candidates,
df.abs = cabs,
algorithm = lima,
n = 0,
seed.number = 123,
check.validity = TRUE)
#> $age
#>  [1] NA NA NA NA NA NA NA NA NA NA
#> 
#> $dialysis
#>  [1] NA NA NA NA NA NA NA NA NA NA
#> 
#> $cPRA
#>  [1] NA NA NA NA NA NA NA NA NA NA
#> 
#> $mmHLA
#> $mmHLA[[1]]
#> 
#>  0  1  2  3  4  5  6 
#>  1  7 19 57 31 22  2 
#> 
#> $mmHLA[[2]]
#> 
#>  1  2  3  4  5  6 
#>  5 26 52 35 16  5 
#> 
#> $mmHLA[[3]]
#> 
#>  0  1  2  3  4  5  6 
#>  1  6 21 59 32 13  7 
#> 
#> $mmHLA[[4]]
#> 
#>  1  2  3  4  5  6 
#>  4 24 51 35 21  4 
#> 
#> $mmHLA[[5]]
#> 
#>  0  1  2  3  4  5  6 
#>  1  7 22 55 31 20  3 
#> 
#> $mmHLA[[6]]
#> 
#>  0  1  2  3  4  5  6 
#>  1  6 22 57 31 16  6 
#> 
#> $mmHLA[[7]]
#> 
#>  1  2  3  4  5  6 
#>  5 28 51 33 18  4 
#> 
#> $mmHLA[[8]]
#> 
#>  1  2  3  4  5  6 
#>  5 24 58 31 14  7 
#> 
#> $mmHLA[[9]]
#> 
#>  0  1  2  3  4  5  6 
#>  1  8 22 51 34 18  5 
#> 
#> $mmHLA[[10]]
#> 
#>  1  2  3  4  5  6 
#>  7 20 58 30 19  5 
#> 
#> 
#> $mmA
#> $mmA[[1]]
#> 
#>  0  1  2 
#> 22 89 28 
#> 
#> $mmA[[2]]
#> 
#>  0  1  2 
#> 26 82 31 
#> 
#> $mmA[[3]]
#> 
#>  0  1  2 
#> 25 84 30 
#> 
#> $mmA[[4]]
#> 
#>  0  1  2 
#> 24 87 28 
#> 
#> $mmA[[5]]
#> 
#>  0  1  2 
#> 29 82 28 
#> 
#> $mmA[[6]]
#> 
#>  0  1  2 
#> 29 83 27 
#> 
#> $mmA[[7]]
#> 
#>  0  1  2 
#> 23 90 26 
#> 
#> $mmA[[8]]
#> 
#>  0  1  2 
#> 28 84 27 
#> 
#> $mmA[[9]]
#> 
#>  0  1  2 
#> 35 74 30 
#> 
#> $mmA[[10]]
#> 
#>  0  1  2 
#> 24 90 25 
#> 
#> 
#> $mmB
#> $mmB[[1]]
#> 
#>  0  1  2 
#> 16 72 51 
#> 
#> $mmB[[2]]
#> 
#>  0  1  2 
#> 15 74 50 
#> 
#> $mmB[[3]]
#> 
#>  0  1  2 
#> 14 79 46 
#> 
#> $mmB[[4]]
#> 
#>  0  1  2 
#> 13 77 49 
#> 
#> $mmB[[5]]
#> 
#>  0  1  2 
#> 19 70 50 
#> 
#> $mmB[[6]]
#> 
#>  0  1  2 
#> 14 75 50 
#> 
#> $mmB[[7]]
#> 
#>  0  1  2 
#> 17 73 49 
#> 
#> $mmB[[8]]
#> 
#>  0  1  2 
#> 13 75 51 
#> 
#> $mmB[[9]]
#> 
#>  0  1  2 
#> 14 71 54 
#> 
#> $mmB[[10]]
#> 
#>  0  1  2 
#> 14 72 53 
#> 
#> 
#> $mmDR
#> $mmDR[[1]]
#> 
#>  0  1  2 
#> 29 77 33 
#> 
#> $mmDR[[2]]
#> 
#>  0  1  2 
#> 21 91 27 
#> 
#> $mmDR[[3]]
#> 
#>  0  1  2 
#> 24 85 30 
#> 
#> $mmDR[[4]]
#> 
#>  0  1  2 
#> 22 78 39 
#> 
#> $mmDR[[5]]
#> 
#>  0  1  2 
#> 23 82 34 
#> 
#> $mmDR[[6]]
#> 
#>  0  1  2 
#> 24 81 34 
#> 
#> $mmDR[[7]]
#> 
#>  0  1  2 
#> 26 79 34 
#> 
#> $mmDR[[8]]
#> 
#>  0  1  2 
#> 21 88 30 
#> 
#> $mmDR[[9]]
#> 
#>  0  1  2 
#> 26 78 35 
#> 
#> $mmDR[[10]]
#> 
#>  0  1  2 
#> 22 86 31 
#> 
#> 
#> $ABO
#> $ABO[[1]]
#> 
#>  A AB  B  O 
#> 64  5 12 58 
#> 
#> $ABO[[2]]
#> 
#>  A AB  B  O 
#> 64  5 12 58 
#> 
#> $ABO[[3]]
#> 
#>  A AB  B  O 
#> 64  5 12 58 
#> 
#> $ABO[[4]]
#> 
#>  A AB  B  O 
#> 64  5 12 58 
#> 
#> $ABO[[5]]
#> 
#>  A AB  B  O 
#> 64  5 12 58 
#> 
#> $ABO[[6]]
#> 
#>  A AB  B  O 
#> 64  5 12 58 
#> 
#> $ABO[[7]]
#> 
#>  A AB  B  O 
#> 64  5 12 58 
#> 
#> $ABO[[8]]
#> 
#>  A AB  B  O 
#> 64  5 12 58 
#> 
#> $ABO[[9]]
#> 
#>  A AB  B  O 
#> 64  5 12 58 
#> 
#> $ABO[[10]]
#> 
#>  A AB  B  O 
#> 64  5 12 58 
#> 
#> 
#> $HI
#> $HI[[1]]
#> 
#> FALSE  TRUE 
#>   122    17 
#> 
#> $HI[[2]]
#> 
#> FALSE  TRUE 
#>   122    17 
#> 
#> $HI[[3]]
#> 
#> FALSE  TRUE 
#>   123    16 
#> 
#> $HI[[4]]
#> 
#> FALSE  TRUE 
#>   122    17 
#> 
#> $HI[[5]]
#> 
#> FALSE  TRUE 
#>   122    17 
#> 
#> $HI[[6]]
#> 
#> FALSE  TRUE 
#>   122    17 
#> 
#> $HI[[7]]
#> 
#> FALSE  TRUE 
#>   122    17 
#> 
#> $HI[[8]]
#> 
#> FALSE  TRUE 
#>   122    17 
#> 
#> $HI[[9]]
#> 
#> FALSE  TRUE 
#>   122    17 
#> 
#> $HI[[10]]
#> 
#> FALSE  TRUE 
#>   122    17 
#> 
#> 
#> $color
#> $color[[1]]
#> 
#>    Red Orange Yellow  Green 
#>      0     21     97     21 
#> 
#> $color[[2]]
#> 
#>    Red Orange Yellow  Green 
#>      0     21     97     21 
#> 
#> $color[[3]]
#> 
#>    Red Orange Yellow  Green 
#>      0     20     98     21 
#> 
#> $color[[4]]
#> 
#>    Red Orange Yellow  Green 
#>      0     21     98     20 
#> 
#> $color[[5]]
#> 
#>    Red Orange Yellow  Green 
#>      0     21     98     20 
#> 
#> $color[[6]]
#> 
#>    Red Orange Yellow  Green 
#>      0     21     97     21 
#> 
#> $color[[7]]
#> 
#>    Red Orange Yellow  Green 
#>      0     21     97     21 
#> 
#> $color[[8]]
#> 
#>    Red Orange Yellow  Green 
#>      0     21     99     19 
#> 
#> $color[[9]]
#> 
#>    Red Orange Yellow  Green 
#>      0     21     99     19 
#> 
#> $color[[10]]
#> 
#>    Red Orange Yellow  Green 
#>      0     21     97     21 
#> 
#> 
#> $SP
#> $SP[[1]]
#> 
#>   1   2 
#>  19 120 
#> 
#> $SP[[2]]
#> 
#>   1   2 
#>  19 120 
#> 
#> $SP[[3]]
#> 
#>   1   2 
#>  19 120 
#> 
#> $SP[[4]]
#> 
#>   1   2 
#>  19 120 
#> 
#> $SP[[5]]
#> 
#>   1   2 
#>  19 120 
#> 
#> $SP[[6]]
#> 
#>   1   2 
#>  19 120 
#> 
#> $SP[[7]]
#> 
#>   1   2 
#>  19 120 
#> 
#> $SP[[8]]
#> 
#>   1   2 
#>  19 120 
#> 
#> $SP[[9]]
#> 
#>   1   2 
#>  19 120 
#> 
#> $SP[[10]]
#> 
#>   1   2 
#>  19 120 
#> 
#>