Runs several time the function donor_recipient_pairs() as bootstrap.

Generic function that runs the matchability between all combinations of donors and candidates. Runs an arbitrary number of times (iteration.number) to provide statistics.

Usage

several(
  iteration.number = 10,
  df.donors = donors,
  df.candidates = candidates,
  df.abs = cabs,
  algorithm = lima,
  n = 0,
  seed.number = 123,
  check.validity = TRUE,
  ...
)

Arguments

iteration.number

Number of times the matchability runs.

df.donors

A data frame containing demographics and medical information for a pool of donors. For uk algorithm must have their respective columns.

df.candidates

A data frame containing demographics and medical information for a group of waitlisted transplant candidates. For uk algorithm must have respective columns.

df.abs

A data frame with candidates' antibodies.

algorithm

The name of the function to use. Valid options are: lima, et, pts, uk (without quotation)

n

A positive integer to slice the first candidates.

seed.number

Seed for new random number. seed.number can be NA in which case no seed is applied.

check.validity

Logical to decide whether to validate input.

...

all the parameters used on the function algorithm

Value

Overall statistics obtained from all runs.

Examples

# \donttest{
several(iteration.number = 10,
df.donors = donors,
df.candidates = candidates,
df.abs = cabs,
algorithm = lima,
n = 0,
seed.number = 123,
check.validity = TRUE)
#> $age
#>    [1] 0 1 1 0 1 2 1 2 2 2 1 2 2 1 0 0 1 1 0 2 2 1 1 2 2 2 1 2 1 2 2 1 2 1 2 2 2
#>   [38] 1 1 1 0 0 2 1 2 2 1 2 1 2 2 2 2 1 2 2 1 1 1 2 1 2 1 1 0 1 1 2 1 1 1 1 1 1
#>   [75] 1 1 1 0 1 1 1 0 1 2 2 1 1 1 2 2 2 1 1 2 1 0 1 2 1 1 2 2 1 2 1 2 2 0 1 1 2
#>  [112] 2 2 2 1 2 2 0 1 2 1 2 1 1 2 0 2 2 1 1 0 2 2 1 2 1 0 0 0 1 1 2 2 2 2 0 2 2
#>  [149] 0 1 2 1 1 1 2 0 1 2 1 2 1 1 0 1 0 1 1 1 1 1 2 1 1 1 1 1 0 0 1 2 2 0 0 0 2
#>  [186] 2 2 1 2 1 2 1 1 2 1 1 2 2 1 2 0 0 1 1 1 2 0 0 2 1 1 1 1 0 2 0 0 2 1 1 1 1
#>  [223] 1 1 1 2 1 2 1 2 2 2 2 1 1 2 2 2 2 1 1 1 2 2 1 1 1 2 1 1 1 1 2 2 2 1 2 2 1
#>  [260] 1 2 2 1 1 1 2 1 0 2 1 2 2 1 1 2 1 1 1 1 2 1 1 1 2 1 1 1 1 1 2 1 2 0 0 1 2
#>  [297] 2 0 0 0 1 1 0 2 1 1 2 2 1 1 2 2 2 2 1 1 1 2 1 1 1 1 1 1 1 2 0 0 2 1 1 1 2
#>  [334] 1 1 1 1 2 1 1 1 1 1 1 2 2 1 2 2 1 2 1 2 1 1 1 0 2 2 0 1 1 1 1 1 2 1 1 0 2
#>  [371] 1 2 1 1 1 2 2 2 2 2 2 2 2 2 1 1 1 1 1 1 1 2 1 2 2 1 1 1 1 2 2 1 2 1 1 1 1
#>  [408] 0 2 1 2 2 2 1 1 1 1 1 0 0 2 1 1 1 1 2 2 1 1 1 1 2 2 1 1 2 2 1 2 2 1 1 2 1
#>  [445] 1 0 1 1 1 2 1 1 1 2 1 1 1 1 1 1 1 1 2 0 2 2 1 1 2 1 1 2 2 1 1 2 2 1 2 1 2
#>  [482] 1 2 1 2 1 2 2 1 2 1 1 1 2 0 1 1 1 1 2 2 1 2 2 1 2 1 0 2 1 1 1 1 1 2 1 1 2
#>  [519] 1 1 1 0 1 1 1 1 1 1 0 2 1 1 1 1 2 2 0 0 1 0 2 0 1 2 1 0 1 1 2 1 2 2 0 0 2
#>  [556] 1 1 1 1 0 1 1 2 1 1 1 2 0 1 2 1 1 0 0 1 1 1 1 1 1 1 2 0 1 2 1 1 1 1 1 1 1
#>  [593] 2 2 1 1 2 2 1 1 0 0 0 0 0 2 1 1 1 1 1 0 1 1 2 2 0 1 1 0 1 1 1 2 2 2 2 2 1
#>  [630] 1 1 1 1 1 1 2 2 0 2 0 2 2 1 1 1 1 1 1 1 2 2 2 2 1 2 1 1 2 1 1 1 1 2 2 2 1
#>  [667] 2 2 2 2 1 2 2 1 1 2 1 1 1 1 2 1 2 2 2 0 1 1 2 1 0 2 0 0 1 0 1 2 1 1 1 2 2
#>  [704] 2 1 1 1 1 1 2 2 2 1 2 2 1 1 1 0 0 1 2 1 0 1 2 1 1 1 1 2 0 0 1 2 1 1 1 2 2
#>  [741] 1 1 2 2 0 0 1 1 1 1 1 1 1 0 1 1 1 1 0 2 1 2 1 1 2 2 1 1 2 0 1 2 1 1 1 2 1
#>  [778] 1 1 1 0 0 1 1 2 2 2 2 1 1 2 1 1 0 0 1 2 1 2 2 0 2 2 1 1 2 1 1 1 1 2 1 0 1
#>  [815] 2 2 2 2 2 2 1 2 1 2 2 2 1 0 0 1 2 0 2 1 1 0 1 0 2 2 2 0 2 2 0 1 2 2 2 0 2
#>  [852] 2 2 0 2 1 1 1 1 1 1 1 2 1 1 1 2 2 1 2 0 1 1 2 1 1 2 1 1 2 1 2 1 2 1 0 0 0
#>  [889] 1 1 1 1 1 1 1 1 0 1 1 1 1 1 2 1 1 2 2 2 1 1 2 1 2 0 1 1 1 1 2 2 1 1 2 1 2
#>  [926] 1 1 2 2 1 1 1 2 1 1 2 1 2 1 2 0 0 1 2 1 2 0 0 1 1 2 0 1 2 1 1 2 1 1 1 1 1
#>  [963] 1 2 2 2 1 1 2 1 2 1 1 0 0 0 1 2 1 0 1 1 2 2 1 2 2 1 1 2 2 1 0 1 1 2 2 1 1
#> [1000] 2 0 0 2 1 1 1 2 1 1 2 1 1 0 0 1 1 2 2 1 1 1 2 1 0 2 0 0 1 2 2 1 2 1 1 1 1
#> [1037] 2 2 2 0 1 2 0 1 1 2 1 1 1 1 2 1 1 1 1 1 1 1 2 1 2 2 1 1 2 2 1 2 2 2 1 1 2
#> [1074] 2 1 1 2 1 2 2 2 2 1 0 2 2 1 1 1 1 1 2 1 1 0 0 1 1 1 0 1 1 2 2 2 0 1 1 2 2
#> [1111] 0 0 1 2 1 1 1 2 1 1 1 2 1 2 2 2 1 1 1 2 0 0 1 2 1 1 2 1 1 1 1 1 1 1 1 1 0
#> [1148] 0 2 1 1 1 2 1 2 2 1 0 1 1 2 1 1 1 2 2 2 1 1 2 2 2 0 0 1 1 1 1 2 1 1 1 2 0
#> [1185] 2 2 2 2 1 2 1 1 0 0 1 1 2 1 2 2 1 1 2 2 1 1 1 2 2 2 2 1 1 1 2 2 1 1 1 1 2
#> [1222] 0 2 2 2 2 1 1 2 1 1 2 2 2 1 2 1 1 2 1 1 1 1 2 1 0 1 0 1 1 1 2 0 2 1 0 1 2
#> [1259] 1 2 1 1 1 0 1 1 2 2 2 1 2 0 1 1 2 2 2 1 2 2 1 1 0 0 1 2 2 1 1 1 2 1 2 2 1
#> [1296] 1 0 1 1 1 1 1 2 1 1 2 1 1 0 0 1 1 2 1 2 1 1 1 1 2 1 2 2 1 1 1 0 1 1 1 0 0
#> [1333] 2 2 2 1 1 0 1 2 1 1 1 1 1 2 1 1 1 2 2 1 2 1 1 1 2 2 1 2 1 2 2 2 2 2 1 2 2
#> [1370] 2 2 2 2 1 1 0 1 1 1 1 1 1 1 1 2 2 1 1 2 1 1 1 1 1 1 1 1 2 1 2
#> 
#> $dialysis
#>    [1] 0 1 1 0 1 2 1 2 2 2 1 2 2 1 0 0 1 1 0 2 2 1 1 2 2 2 1 2 1 2 2 1 2 1 2 2 2
#>   [38] 1 1 1 0 0 2 1 2 2 1 2 1 2 2 2 2 1 2 2 1 1 1 2 1 2 1 1 0 1 1 2 1 1 1 1 1 1
#>   [75] 1 1 1 0 1 1 1 0 1 2 2 1 1 1 2 2 2 1 1 2 1 0 1 2 1 1 2 2 1 2 1 2 2 0 1 1 2
#>  [112] 2 2 2 1 2 2 0 1 2 1 2 1 1 2 0 2 2 1 1 0 2 2 1 2 1 0 0 0 1 1 2 2 2 2 0 2 2
#>  [149] 0 1 2 1 1 1 2 0 1 2 1 2 1 1 0 1 0 1 1 1 1 1 2 1 1 1 1 1 0 0 1 2 2 0 0 0 2
#>  [186] 2 2 1 2 1 2 1 1 2 1 1 2 2 1 2 0 0 1 1 1 2 0 0 2 1 1 1 1 0 2 0 0 2 1 1 1 1
#>  [223] 1 1 1 2 1 2 1 2 2 2 2 1 1 2 2 2 2 1 1 1 2 2 1 1 1 2 1 1 1 1 2 2 2 1 2 2 1
#>  [260] 1 2 2 1 1 1 2 1 0 2 1 2 2 1 1 2 1 1 1 1 2 1 1 1 2 1 1 1 1 1 2 1 2 0 0 1 2
#>  [297] 2 0 0 0 1 1 0 2 1 1 2 2 1 1 2 2 2 2 1 1 1 2 1 1 1 1 1 1 1 2 0 0 2 1 1 1 2
#>  [334] 1 1 1 1 2 1 1 1 1 1 1 2 2 1 2 2 1 2 1 2 1 1 1 0 2 2 0 1 1 1 1 1 2 1 1 0 2
#>  [371] 1 2 1 1 1 2 2 2 2 2 2 2 2 2 1 1 1 1 1 1 1 2 1 2 2 1 1 1 1 2 2 1 2 1 1 1 1
#>  [408] 0 2 1 2 2 2 1 1 1 1 1 0 0 2 1 1 1 1 2 2 1 1 1 1 2 2 1 1 2 2 1 2 2 1 1 2 1
#>  [445] 1 0 1 1 1 2 1 1 1 2 1 1 1 1 1 1 1 1 2 0 2 2 1 1 2 1 1 2 2 1 1 2 2 1 2 1 2
#>  [482] 1 2 1 2 1 2 2 1 2 1 1 1 2 0 1 1 1 1 2 2 1 2 2 1 2 1 0 2 1 1 1 1 1 2 1 1 2
#>  [519] 1 1 1 0 1 1 1 1 1 1 0 2 1 1 1 1 2 2 0 0 1 0 2 0 1 2 1 0 1 1 2 1 2 2 0 0 2
#>  [556] 1 1 1 1 0 1 1 2 1 1 1 2 0 1 2 1 1 0 0 1 1 1 1 1 1 1 2 0 1 2 1 1 1 1 1 1 1
#>  [593] 2 2 1 1 2 2 1 1 0 0 0 0 0 2 1 1 1 1 1 0 1 1 2 2 0 1 1 0 1 1 1 2 2 2 2 2 1
#>  [630] 1 1 1 1 1 1 2 2 0 2 0 2 2 1 1 1 1 1 1 1 2 2 2 2 1 2 1 1 2 1 1 1 1 2 2 2 1
#>  [667] 2 2 2 2 1 2 2 1 1 2 1 1 1 1 2 1 2 2 2 0 1 1 2 1 0 2 0 0 1 0 1 2 1 1 1 2 2
#>  [704] 2 1 1 1 1 1 2 2 2 1 2 2 1 1 1 0 0 1 2 1 0 1 2 1 1 1 1 2 0 0 1 2 1 1 1 2 2
#>  [741] 1 1 2 2 0 0 1 1 1 1 1 1 1 0 1 1 1 1 0 2 1 2 1 1 2 2 1 1 2 0 1 2 1 1 1 2 1
#>  [778] 1 1 1 0 0 1 1 2 2 2 2 1 1 2 1 1 0 0 1 2 1 2 2 0 2 2 1 1 2 1 1 1 1 2 1 0 1
#>  [815] 2 2 2 2 2 2 1 2 1 2 2 2 1 0 0 1 2 0 2 1 1 0 1 0 2 2 2 0 2 2 0 1 2 2 2 0 2
#>  [852] 2 2 0 2 1 1 1 1 1 1 1 2 1 1 1 2 2 1 2 0 1 1 2 1 1 2 1 1 2 1 2 1 2 1 0 0 0
#>  [889] 1 1 1 1 1 1 1 1 0 1 1 1 1 1 2 1 1 2 2 2 1 1 2 1 2 0 1 1 1 1 2 2 1 1 2 1 2
#>  [926] 1 1 2 2 1 1 1 2 1 1 2 1 2 1 2 0 0 1 2 1 2 0 0 1 1 2 0 1 2 1 1 2 1 1 1 1 1
#>  [963] 1 2 2 2 1 1 2 1 2 1 1 0 0 0 1 2 1 0 1 1 2 2 1 2 2 1 1 2 2 1 0 1 1 2 2 1 1
#> [1000] 2 0 0 2 1 1 1 2 1 1 2 1 1 0 0 1 1 2 2 1 1 1 2 1 0 2 0 0 1 2 2 1 2 1 1 1 1
#> [1037] 2 2 2 0 1 2 0 1 1 2 1 1 1 1 2 1 1 1 1 1 1 1 2 1 2 2 1 1 2 2 1 2 2 2 1 1 2
#> [1074] 2 1 1 2 1 2 2 2 2 1 0 2 2 1 1 1 1 1 2 1 1 0 0 1 1 1 0 1 1 2 2 2 0 1 1 2 2
#> [1111] 0 0 1 2 1 1 1 2 1 1 1 2 1 2 2 2 1 1 1 2 0 0 1 2 1 1 2 1 1 1 1 1 1 1 1 1 0
#> [1148] 0 2 1 1 1 2 1 2 2 1 0 1 1 2 1 1 1 2 2 2 1 1 2 2 2 0 0 1 1 1 1 2 1 1 1 2 0
#> [1185] 2 2 2 2 1 2 1 1 0 0 1 1 2 1 2 2 1 1 2 2 1 1 1 2 2 2 2 1 1 1 2 2 1 1 1 1 2
#> [1222] 0 2 2 2 2 1 1 2 1 1 2 2 2 1 2 1 1 2 1 1 1 1 2 1 0 1 0 1 1 1 2 0 2 1 0 1 2
#> [1259] 1 2 1 1 1 0 1 1 2 2 2 1 2 0 1 1 2 2 2 1 2 2 1 1 0 0 1 2 2 1 1 1 2 1 2 2 1
#> [1296] 1 0 1 1 1 1 1 2 1 1 2 1 1 0 0 1 1 2 1 2 1 1 1 1 2 1 2 2 1 1 1 0 1 1 1 0 0
#> [1333] 2 2 2 1 1 0 1 2 1 1 1 1 1 2 1 1 1 2 2 1 2 1 1 1 2 2 1 2 1 2 2 2 2 2 1 2 2
#> [1370] 2 2 2 2 1 1 0 1 1 1 1 1 1 1 1 2 2 1 1 2 1 1 1 1 1 1 1 1 2 1 2
#> 
#> $cPRA
#>    [1] 0 1 1 0 1 2 1 2 2 2 1 2 2 1 0 0 1 1 0 2 2 1 1 2 2 2 1 2 1 2 2 1 2 1 2 2 2
#>   [38] 1 1 1 0 0 2 1 2 2 1 2 1 2 2 2 2 1 2 2 1 1 1 2 1 2 1 1 0 1 1 2 1 1 1 1 1 1
#>   [75] 1 1 1 0 1 1 1 0 1 2 2 1 1 1 2 2 2 1 1 2 1 0 1 2 1 1 2 2 1 2 1 2 2 0 1 1 2
#>  [112] 2 2 2 1 2 2 0 1 2 1 2 1 1 2 0 2 2 1 1 0 2 2 1 2 1 0 0 0 1 1 2 2 2 2 0 2 2
#>  [149] 0 1 2 1 1 1 2 0 1 2 1 2 1 1 0 1 0 1 1 1 1 1 2 1 1 1 1 1 0 0 1 2 2 0 0 0 2
#>  [186] 2 2 1 2 1 2 1 1 2 1 1 2 2 1 2 0 0 1 1 1 2 0 0 2 1 1 1 1 0 2 0 0 2 1 1 1 1
#>  [223] 1 1 1 2 1 2 1 2 2 2 2 1 1 2 2 2 2 1 1 1 2 2 1 1 1 2 1 1 1 1 2 2 2 1 2 2 1
#>  [260] 1 2 2 1 1 1 2 1 0 2 1 2 2 1 1 2 1 1 1 1 2 1 1 1 2 1 1 1 1 1 2 1 2 0 0 1 2
#>  [297] 2 0 0 0 1 1 0 2 1 1 2 2 1 1 2 2 2 2 1 1 1 2 1 1 1 1 1 1 1 2 0 0 2 1 1 1 2
#>  [334] 1 1 1 1 2 1 1 1 1 1 1 2 2 1 2 2 1 2 1 2 1 1 1 0 2 2 0 1 1 1 1 1 2 1 1 0 2
#>  [371] 1 2 1 1 1 2 2 2 2 2 2 2 2 2 1 1 1 1 1 1 1 2 1 2 2 1 1 1 1 2 2 1 2 1 1 1 1
#>  [408] 0 2 1 2 2 2 1 1 1 1 1 0 0 2 1 1 1 1 2 2 1 1 1 1 2 2 1 1 2 2 1 2 2 1 1 2 1
#>  [445] 1 0 1 1 1 2 1 1 1 2 1 1 1 1 1 1 1 1 2 0 2 2 1 1 2 1 1 2 2 1 1 2 2 1 2 1 2
#>  [482] 1 2 1 2 1 2 2 1 2 1 1 1 2 0 1 1 1 1 2 2 1 2 2 1 2 1 0 2 1 1 1 1 1 2 1 1 2
#>  [519] 1 1 1 0 1 1 1 1 1 1 0 2 1 1 1 1 2 2 0 0 1 0 2 0 1 2 1 0 1 1 2 1 2 2 0 0 2
#>  [556] 1 1 1 1 0 1 1 2 1 1 1 2 0 1 2 1 1 0 0 1 1 1 1 1 1 1 2 0 1 2 1 1 1 1 1 1 1
#>  [593] 2 2 1 1 2 2 1 1 0 0 0 0 0 2 1 1 1 1 1 0 1 1 2 2 0 1 1 0 1 1 1 2 2 2 2 2 1
#>  [630] 1 1 1 1 1 1 2 2 0 2 0 2 2 1 1 1 1 1 1 1 2 2 2 2 1 2 1 1 2 1 1 1 1 2 2 2 1
#>  [667] 2 2 2 2 1 2 2 1 1 2 1 1 1 1 2 1 2 2 2 0 1 1 2 1 0 2 0 0 1 0 1 2 1 1 1 2 2
#>  [704] 2 1 1 1 1 1 2 2 2 1 2 2 1 1 1 0 0 1 2 1 0 1 2 1 1 1 1 2 0 0 1 2 1 1 1 2 2
#>  [741] 1 1 2 2 0 0 1 1 1 1 1 1 1 0 1 1 1 1 0 2 1 2 1 1 2 2 1 1 2 0 1 2 1 1 1 2 1
#>  [778] 1 1 1 0 0 1 1 2 2 2 2 1 1 2 1 1 0 0 1 2 1 2 2 0 2 2 1 1 2 1 1 1 1 2 1 0 1
#>  [815] 2 2 2 2 2 2 1 2 1 2 2 2 1 0 0 1 2 0 2 1 1 0 1 0 2 2 2 0 2 2 0 1 2 2 2 0 2
#>  [852] 2 2 0 2 1 1 1 1 1 1 1 2 1 1 1 2 2 1 2 0 1 1 2 1 1 2 1 1 2 1 2 1 2 1 0 0 0
#>  [889] 1 1 1 1 1 1 1 1 0 1 1 1 1 1 2 1 1 2 2 2 1 1 2 1 2 0 1 1 1 1 2 2 1 1 2 1 2
#>  [926] 1 1 2 2 1 1 1 2 1 1 2 1 2 1 2 0 0 1 2 1 2 0 0 1 1 2 0 1 2 1 1 2 1 1 1 1 1
#>  [963] 1 2 2 2 1 1 2 1 2 1 1 0 0 0 1 2 1 0 1 1 2 2 1 2 2 1 1 2 2 1 0 1 1 2 2 1 1
#> [1000] 2 0 0 2 1 1 1 2 1 1 2 1 1 0 0 1 1 2 2 1 1 1 2 1 0 2 0 0 1 2 2 1 2 1 1 1 1
#> [1037] 2 2 2 0 1 2 0 1 1 2 1 1 1 1 2 1 1 1 1 1 1 1 2 1 2 2 1 1 2 2 1 2 2 2 1 1 2
#> [1074] 2 1 1 2 1 2 2 2 2 1 0 2 2 1 1 1 1 1 2 1 1 0 0 1 1 1 0 1 1 2 2 2 0 1 1 2 2
#> [1111] 0 0 1 2 1 1 1 2 1 1 1 2 1 2 2 2 1 1 1 2 0 0 1 2 1 1 2 1 1 1 1 1 1 1 1 1 0
#> [1148] 0 2 1 1 1 2 1 2 2 1 0 1 1 2 1 1 1 2 2 2 1 1 2 2 2 0 0 1 1 1 1 2 1 1 1 2 0
#> [1185] 2 2 2 2 1 2 1 1 0 0 1 1 2 1 2 2 1 1 2 2 1 1 1 2 2 2 2 1 1 1 2 2 1 1 1 1 2
#> [1222] 0 2 2 2 2 1 1 2 1 1 2 2 2 1 2 1 1 2 1 1 1 1 2 1 0 1 0 1 1 1 2 0 2 1 0 1 2
#> [1259] 1 2 1 1 1 0 1 1 2 2 2 1 2 0 1 1 2 2 2 1 2 2 1 1 0 0 1 2 2 1 1 1 2 1 2 2 1
#> [1296] 1 0 1 1 1 1 1 2 1 1 2 1 1 0 0 1 1 2 1 2 1 1 1 1 2 1 2 2 1 1 1 0 1 1 1 0 0
#> [1333] 2 2 2 1 1 0 1 2 1 1 1 1 1 2 1 1 1 2 2 1 2 1 1 1 2 2 1 2 1 2 2 2 2 2 1 2 2
#> [1370] 2 2 2 2 1 1 0 1 1 1 1 1 1 1 1 2 2 1 1 2 1 1 1 1 1 1 1 1 2 1 2
#> 
#> $mmHLA
#> $mmHLA[[1]]
#>   [1] 0 1 1 0 1 2 1 2 2 2 1 2 2 1 0 0 1 1 0 2 2 1 1 2 2 2 1 2 1 2 2 1 2 1 2 2 2
#>  [38] 1 1 1 0 0 2 1 2 2 1 2 1 2 2 2 2 1 2 2 1 1 1 2 1 2 1 1 0 1 1 2 1 1 1 1 1 1
#>  [75] 1 1 1 0 1 1 1 0 1 2 2 1 1 1 2 2 2 1 1 2 1 0 1 2 1 1 2 2 1 2 1 2 2 0 1 1 2
#> [112] 2 2 2 1 2 2 0 1 2 1 2 1 1 2 0 2 2 1 1 0 2 2 1 2 1 0 0 0 1
#> 
#> $mmHLA[[2]]
#>   [1] 1 2 2 2 2 0 2 2 0 1 2 1 1 1 2 0 1 2 1 2 1 1 0 1 0 1 1 1 1 1 2 1 1 1 1 1 0
#>  [38] 0 1 2 2 0 0 0 2 2 2 1 2 1 2 1 1 2 1 1 2 2 1 2 0 0 1 1 1 2 0 0 2 1 1 1 1 0
#>  [75] 2 0 0 2 1 1 1 1 1 1 1 2 1 2 1 2 2 2 2 1 1 2 2 2 2 1 1 1 2 2 1 1 1 2 1 1 1
#> [112] 1 2 2 2 1 2 2 1 1 2 2 1 1 1 2 1 0 2 1 2 2 1 1 2 1 1 1 1 2
#> 
#> $mmHLA[[3]]
#>   [1] 1 1 1 2 1 1 1 1 1 2 1 2 0 0 1 2 2 0 0 0 1 1 0 2 1 1 2 2 1 1 2 2 2 2 1 1 1
#>  [38] 2 1 1 1 1 1 1 1 2 0 0 2 1 1 1 2 1 1 1 1 2 1 1 1 1 1 1 2 2 1 2 2 1 2 1 2 1
#>  [75] 1 1 0 2 2 0 1 1 1 1 1 2 1 1 0 2 1 2 1 1 1 2 2 2 2 2 2 2 2 2 1 1 1 1 1 1 1
#> [112] 2 1 2 2 1 1 1 1 2 2 1 2 1 1 1 1 0 2 1 2 2 2 1 1 1 1 1 0 0
#> 
#> $mmHLA[[4]]
#>   [1] 2 1 1 1 1 2 2 1 1 1 1 2 2 1 1 2 2 1 2 2 1 1 2 1 1 0 1 1 1 2 1 1 1 2 1 1 1
#>  [38] 1 1 1 1 1 2 0 2 2 1 1 2 1 1 2 2 1 1 2 2 1 2 1 2 1 2 1 2 1 2 2 1 2 1 1 1 2
#>  [75] 0 1 1 1 1 2 2 1 2 2 1 2 1 0 2 1 1 1 1 1 2 1 1 2 1 1 1 0 1 1 1 1 1 1 0 2 1
#> [112] 1 1 1 2 2 0 0 1 0 2 0 1 2 1 0 1 1 2 1 2 2 0 0 2 1 1 1 1 0
#> 
#> $mmHLA[[5]]
#>   [1] 1 1 2 1 1 1 2 0 1 2 1 1 0 0 1 1 1 1 1 1 1 2 0 1 2 1 1 1 1 1 1 1 2 2 1 1 2
#>  [38] 2 1 1 0 0 0 0 0 2 1 1 1 1 1 0 1 1 2 2 0 1 1 0 1 1 1 2 2 2 2 2 1 1 1 1 1 1
#>  [75] 1 2 2 0 2 0 2 2 1 1 1 1 1 1 1 2 2 2 2 1 2 1 1 2 1 1 1 1 2 2 2 1 2 2 2 2 1
#> [112] 2 2 1 1 2 1 1 1 1 2 1 2 2 2 0 1 1 2 1 0 2 0 0 1 0 1 2 1 1
#> 
#> $mmHLA[[6]]
#>   [1] 1 2 2 2 1 1 1 1 1 2 2 2 1 2 2 1 1 1 0 0 1 2 1 0 1 2 1 1 1 1 2 0 0 1 2 1 1
#>  [38] 1 2 2 1 1 2 2 0 0 1 1 1 1 1 1 1 0 1 1 1 1 0 2 1 2 1 1 2 2 1 1 2 0 1 2 1 1
#>  [75] 1 2 1 1 1 1 0 0 1 1 2 2 2 2 1 1 2 1 1 0 0 1 2 1 2 2 0 2 2 1 1 2 1 1 1 1 2
#> [112] 1 0 1 2 2 2 2 2 2 1 2 1 2 2 2 1 0 0 1 2 0 2 1 1 0 1 0 2 2
#> 
#> $mmHLA[[7]]
#>   [1] 2 0 2 2 0 1 2 2 2 0 2 2 2 0 2 1 1 1 1 1 1 1 2 1 1 1 2 2 1 2 0 1 1 2 1 1 2
#>  [38] 1 1 2 1 2 1 2 1 0 0 0 1 1 1 1 1 1 1 1 0 1 1 1 1 1 2 1 1 2 2 2 1 1 2 1 2 0
#>  [75] 1 1 1 1 2 2 1 1 2 1 2 1 1 2 2 1 1 1 2 1 1 2 1 2 1 2 0 0 1 2 1 2 0 0 1 1 2
#> [112] 0 1 2 1 1 2 1 1 1 1 1 1 2 2 2 1 1 2 1 2 1 1 0 0 0 1 2 1 0
#> 
#> $mmHLA[[8]]
#>   [1] 1 1 2 2 1 2 2 1 1 2 2 1 0 1 1 2 2 1 1 2 0 0 2 1 1 1 2 1 1 2 1 1 0 0 1 1 2
#>  [38] 2 1 1 1 2 1 0 2 0 0 1 2 2 1 2 1 1 1 1 2 2 2 0 1 2 0 1 1 2 1 1 1 1 2 1 1 1
#>  [75] 1 1 1 1 2 1 2 2 1 1 2 2 1 2 2 2 1 1 2 2 1 1 2 1 2 2 2 2 1 0 2 2 1 1 1 1 1
#> [112] 2 1 1 0 0 1 1 1 0 1 1 2 2 2 0 1 1 2 2 0 0 1 2 1 1 1 2 1 1
#> 
#> $mmHLA[[9]]
#>   [1] 1 2 1 2 2 2 1 1 1 2 0 0 1 2 1 1 2 1 1 1 1 1 1 1 1 1 0 0 2 1 1 1 2 1 2 2 1
#>  [38] 0 1 1 2 1 1 1 2 2 2 1 1 2 2 2 0 0 1 1 1 1 2 1 1 1 2 0 2 2 2 2 1 2 1 1 0 0
#>  [75] 1 1 2 1 2 2 1 1 2 2 1 1 1 2 2 2 2 1 1 1 2 2 1 1 1 1 2 0 2 2 2 2 1 1 2 1 1
#> [112] 2 2 2 1 2 1 1 2 1 1 1 1 2 1 0 1 0 1 1 1 2 0 2 1 0 1 2 1 2
#> 
#> $mmHLA[[10]]
#>   [1] 1 1 1 0 1 1 2 2 2 1 2 0 1 1 2 2 2 1 2 2 1 1 0 0 1 2 2 1 1 1 2 1 2 2 1 1 0
#>  [38] 1 1 1 1 1 2 1 1 2 1 1 0 0 1 1 2 1 2 1 1 1 1 2 1 2 2 1 1 1 0 1 1 1 0 0 2 2
#>  [75] 2 1 1 0 1 2 1 1 1 1 1 2 1 1 1 2 2 1 2 1 1 1 2 2 1 2 1 2 2 2 2 2 1 2 2 2 2
#> [112] 2 2 1 1 0 1 1 1 1 1 1 1 1 2 2 1 1 2 1 1 1 1 1 1 1 1 2 1 2
#> 
#> 
#> $mmA
#> $mmA[[1]]
#>   [1] 0 1 1 0 1 2 1 2 2 2 1 2 2 1 0 0 1 1 0 2 2 1 1 2 2 2 1 2 1 2 2 1 2 1 2 2 2
#>  [38] 1 1 1 0 0 2 1 2 2 1 2 1 2 2 2 2 1 2 2 1 1 1 2 1 2 1 1 0 1 1 2 1 1 1 1 1 1
#>  [75] 1 1 1 0 1 1 1 0 1 2 2 1 1 1 2 2 2 1 1 2 1 0 1 2 1 1 2 2 1 2 1 2 2 0 1 1 2
#> [112] 2 2 2 1 2 2 0 1 2 1 2 1 1 2 0 2 2 1 1 0 2 2 1 2 1 0 0 0 1
#> 
#> $mmA[[2]]
#>   [1] 1 2 2 2 2 0 2 2 0 1 2 1 1 1 2 0 1 2 1 2 1 1 0 1 0 1 1 1 1 1 2 1 1 1 1 1 0
#>  [38] 0 1 2 2 0 0 0 2 2 2 1 2 1 2 1 1 2 1 1 2 2 1 2 0 0 1 1 1 2 0 0 2 1 1 1 1 0
#>  [75] 2 0 0 2 1 1 1 1 1 1 1 2 1 2 1 2 2 2 2 1 1 2 2 2 2 1 1 1 2 2 1 1 1 2 1 1 1
#> [112] 1 2 2 2 1 2 2 1 1 2 2 1 1 1 2 1 0 2 1 2 2 1 1 2 1 1 1 1 2
#> 
#> $mmA[[3]]
#>   [1] 1 1 1 2 1 1 1 1 1 2 1 2 0 0 1 2 2 0 0 0 1 1 0 2 1 1 2 2 1 1 2 2 2 2 1 1 1
#>  [38] 2 1 1 1 1 1 1 1 2 0 0 2 1 1 1 2 1 1 1 1 2 1 1 1 1 1 1 2 2 1 2 2 1 2 1 2 1
#>  [75] 1 1 0 2 2 0 1 1 1 1 1 2 1 1 0 2 1 2 1 1 1 2 2 2 2 2 2 2 2 2 1 1 1 1 1 1 1
#> [112] 2 1 2 2 1 1 1 1 2 2 1 2 1 1 1 1 0 2 1 2 2 2 1 1 1 1 1 0 0
#> 
#> $mmA[[4]]
#>   [1] 2 1 1 1 1 2 2 1 1 1 1 2 2 1 1 2 2 1 2 2 1 1 2 1 1 0 1 1 1 2 1 1 1 2 1 1 1
#>  [38] 1 1 1 1 1 2 0 2 2 1 1 2 1 1 2 2 1 1 2 2 1 2 1 2 1 2 1 2 1 2 2 1 2 1 1 1 2
#>  [75] 0 1 1 1 1 2 2 1 2 2 1 2 1 0 2 1 1 1 1 1 2 1 1 2 1 1 1 0 1 1 1 1 1 1 0 2 1
#> [112] 1 1 1 2 2 0 0 1 0 2 0 1 2 1 0 1 1 2 1 2 2 0 0 2 1 1 1 1 0
#> 
#> $mmA[[5]]
#>   [1] 1 1 2 1 1 1 2 0 1 2 1 1 0 0 1 1 1 1 1 1 1 2 0 1 2 1 1 1 1 1 1 1 2 2 1 1 2
#>  [38] 2 1 1 0 0 0 0 0 2 1 1 1 1 1 0 1 1 2 2 0 1 1 0 1 1 1 2 2 2 2 2 1 1 1 1 1 1
#>  [75] 1 2 2 0 2 0 2 2 1 1 1 1 1 1 1 2 2 2 2 1 2 1 1 2 1 1 1 1 2 2 2 1 2 2 2 2 1
#> [112] 2 2 1 1 2 1 1 1 1 2 1 2 2 2 0 1 1 2 1 0 2 0 0 1 0 1 2 1 1
#> 
#> $mmA[[6]]
#>   [1] 1 2 2 2 1 1 1 1 1 2 2 2 1 2 2 1 1 1 0 0 1 2 1 0 1 2 1 1 1 1 2 0 0 1 2 1 1
#>  [38] 1 2 2 1 1 2 2 0 0 1 1 1 1 1 1 1 0 1 1 1 1 0 2 1 2 1 1 2 2 1 1 2 0 1 2 1 1
#>  [75] 1 2 1 1 1 1 0 0 1 1 2 2 2 2 1 1 2 1 1 0 0 1 2 1 2 2 0 2 2 1 1 2 1 1 1 1 2
#> [112] 1 0 1 2 2 2 2 2 2 1 2 1 2 2 2 1 0 0 1 2 0 2 1 1 0 1 0 2 2
#> 
#> $mmA[[7]]
#>   [1] 2 0 2 2 0 1 2 2 2 0 2 2 2 0 2 1 1 1 1 1 1 1 2 1 1 1 2 2 1 2 0 1 1 2 1 1 2
#>  [38] 1 1 2 1 2 1 2 1 0 0 0 1 1 1 1 1 1 1 1 0 1 1 1 1 1 2 1 1 2 2 2 1 1 2 1 2 0
#>  [75] 1 1 1 1 2 2 1 1 2 1 2 1 1 2 2 1 1 1 2 1 1 2 1 2 1 2 0 0 1 2 1 2 0 0 1 1 2
#> [112] 0 1 2 1 1 2 1 1 1 1 1 1 2 2 2 1 1 2 1 2 1 1 0 0 0 1 2 1 0
#> 
#> $mmA[[8]]
#>   [1] 1 1 2 2 1 2 2 1 1 2 2 1 0 1 1 2 2 1 1 2 0 0 2 1 1 1 2 1 1 2 1 1 0 0 1 1 2
#>  [38] 2 1 1 1 2 1 0 2 0 0 1 2 2 1 2 1 1 1 1 2 2 2 0 1 2 0 1 1 2 1 1 1 1 2 1 1 1
#>  [75] 1 1 1 1 2 1 2 2 1 1 2 2 1 2 2 2 1 1 2 2 1 1 2 1 2 2 2 2 1 0 2 2 1 1 1 1 1
#> [112] 2 1 1 0 0 1 1 1 0 1 1 2 2 2 0 1 1 2 2 0 0 1 2 1 1 1 2 1 1
#> 
#> $mmA[[9]]
#>   [1] 1 2 1 2 2 2 1 1 1 2 0 0 1 2 1 1 2 1 1 1 1 1 1 1 1 1 0 0 2 1 1 1 2 1 2 2 1
#>  [38] 0 1 1 2 1 1 1 2 2 2 1 1 2 2 2 0 0 1 1 1 1 2 1 1 1 2 0 2 2 2 2 1 2 1 1 0 0
#>  [75] 1 1 2 1 2 2 1 1 2 2 1 1 1 2 2 2 2 1 1 1 2 2 1 1 1 1 2 0 2 2 2 2 1 1 2 1 1
#> [112] 2 2 2 1 2 1 1 2 1 1 1 1 2 1 0 1 0 1 1 1 2 0 2 1 0 1 2 1 2
#> 
#> $mmA[[10]]
#>   [1] 1 1 1 0 1 1 2 2 2 1 2 0 1 1 2 2 2 1 2 2 1 1 0 0 1 2 2 1 1 1 2 1 2 2 1 1 0
#>  [38] 1 1 1 1 1 2 1 1 2 1 1 0 0 1 1 2 1 2 1 1 1 1 2 1 2 2 1 1 1 0 1 1 1 0 0 2 2
#>  [75] 2 1 1 0 1 2 1 1 1 1 1 2 1 1 1 2 2 1 2 1 1 1 2 2 1 2 1 2 2 2 2 2 1 2 2 2 2
#> [112] 2 2 1 1 0 1 1 1 1 1 1 1 1 2 2 1 1 2 1 1 1 1 1 1 1 1 2 1 2
#> 
#> 
#> $mmB
#> $mmB[[1]]
#>   [1] 0 1 1 0 1 2 1 2 2 2 1 2 2 1 0 0 1 1 0 2 2 1 1 2 2 2 1 2 1 2 2 1 2 1 2 2 2
#>  [38] 1 1 1 0 0 2 1 2 2 1 2 1 2 2 2 2 1 2 2 1 1 1 2 1 2 1 1 0 1 1 2 1 1 1 1 1 1
#>  [75] 1 1 1 0 1 1 1 0 1 2 2 1 1 1 2 2 2 1 1 2 1 0 1 2 1 1 2 2 1 2 1 2 2 0 1 1 2
#> [112] 2 2 2 1 2 2 0 1 2 1 2 1 1 2 0 2 2 1 1 0 2 2 1 2 1 0 0 0 1
#> 
#> $mmB[[2]]
#>   [1] 1 2 2 2 2 0 2 2 0 1 2 1 1 1 2 0 1 2 1 2 1 1 0 1 0 1 1 1 1 1 2 1 1 1 1 1 0
#>  [38] 0 1 2 2 0 0 0 2 2 2 1 2 1 2 1 1 2 1 1 2 2 1 2 0 0 1 1 1 2 0 0 2 1 1 1 1 0
#>  [75] 2 0 0 2 1 1 1 1 1 1 1 2 1 2 1 2 2 2 2 1 1 2 2 2 2 1 1 1 2 2 1 1 1 2 1 1 1
#> [112] 1 2 2 2 1 2 2 1 1 2 2 1 1 1 2 1 0 2 1 2 2 1 1 2 1 1 1 1 2
#> 
#> $mmB[[3]]
#>   [1] 1 1 1 2 1 1 1 1 1 2 1 2 0 0 1 2 2 0 0 0 1 1 0 2 1 1 2 2 1 1 2 2 2 2 1 1 1
#>  [38] 2 1 1 1 1 1 1 1 2 0 0 2 1 1 1 2 1 1 1 1 2 1 1 1 1 1 1 2 2 1 2 2 1 2 1 2 1
#>  [75] 1 1 0 2 2 0 1 1 1 1 1 2 1 1 0 2 1 2 1 1 1 2 2 2 2 2 2 2 2 2 1 1 1 1 1 1 1
#> [112] 2 1 2 2 1 1 1 1 2 2 1 2 1 1 1 1 0 2 1 2 2 2 1 1 1 1 1 0 0
#> 
#> $mmB[[4]]
#>   [1] 2 1 1 1 1 2 2 1 1 1 1 2 2 1 1 2 2 1 2 2 1 1 2 1 1 0 1 1 1 2 1 1 1 2 1 1 1
#>  [38] 1 1 1 1 1 2 0 2 2 1 1 2 1 1 2 2 1 1 2 2 1 2 1 2 1 2 1 2 1 2 2 1 2 1 1 1 2
#>  [75] 0 1 1 1 1 2 2 1 2 2 1 2 1 0 2 1 1 1 1 1 2 1 1 2 1 1 1 0 1 1 1 1 1 1 0 2 1
#> [112] 1 1 1 2 2 0 0 1 0 2 0 1 2 1 0 1 1 2 1 2 2 0 0 2 1 1 1 1 0
#> 
#> $mmB[[5]]
#>   [1] 1 1 2 1 1 1 2 0 1 2 1 1 0 0 1 1 1 1 1 1 1 2 0 1 2 1 1 1 1 1 1 1 2 2 1 1 2
#>  [38] 2 1 1 0 0 0 0 0 2 1 1 1 1 1 0 1 1 2 2 0 1 1 0 1 1 1 2 2 2 2 2 1 1 1 1 1 1
#>  [75] 1 2 2 0 2 0 2 2 1 1 1 1 1 1 1 2 2 2 2 1 2 1 1 2 1 1 1 1 2 2 2 1 2 2 2 2 1
#> [112] 2 2 1 1 2 1 1 1 1 2 1 2 2 2 0 1 1 2 1 0 2 0 0 1 0 1 2 1 1
#> 
#> $mmB[[6]]
#>   [1] 1 2 2 2 1 1 1 1 1 2 2 2 1 2 2 1 1 1 0 0 1 2 1 0 1 2 1 1 1 1 2 0 0 1 2 1 1
#>  [38] 1 2 2 1 1 2 2 0 0 1 1 1 1 1 1 1 0 1 1 1 1 0 2 1 2 1 1 2 2 1 1 2 0 1 2 1 1
#>  [75] 1 2 1 1 1 1 0 0 1 1 2 2 2 2 1 1 2 1 1 0 0 1 2 1 2 2 0 2 2 1 1 2 1 1 1 1 2
#> [112] 1 0 1 2 2 2 2 2 2 1 2 1 2 2 2 1 0 0 1 2 0 2 1 1 0 1 0 2 2
#> 
#> $mmB[[7]]
#>   [1] 2 0 2 2 0 1 2 2 2 0 2 2 2 0 2 1 1 1 1 1 1 1 2 1 1 1 2 2 1 2 0 1 1 2 1 1 2
#>  [38] 1 1 2 1 2 1 2 1 0 0 0 1 1 1 1 1 1 1 1 0 1 1 1 1 1 2 1 1 2 2 2 1 1 2 1 2 0
#>  [75] 1 1 1 1 2 2 1 1 2 1 2 1 1 2 2 1 1 1 2 1 1 2 1 2 1 2 0 0 1 2 1 2 0 0 1 1 2
#> [112] 0 1 2 1 1 2 1 1 1 1 1 1 2 2 2 1 1 2 1 2 1 1 0 0 0 1 2 1 0
#> 
#> $mmB[[8]]
#>   [1] 1 1 2 2 1 2 2 1 1 2 2 1 0 1 1 2 2 1 1 2 0 0 2 1 1 1 2 1 1 2 1 1 0 0 1 1 2
#>  [38] 2 1 1 1 2 1 0 2 0 0 1 2 2 1 2 1 1 1 1 2 2 2 0 1 2 0 1 1 2 1 1 1 1 2 1 1 1
#>  [75] 1 1 1 1 2 1 2 2 1 1 2 2 1 2 2 2 1 1 2 2 1 1 2 1 2 2 2 2 1 0 2 2 1 1 1 1 1
#> [112] 2 1 1 0 0 1 1 1 0 1 1 2 2 2 0 1 1 2 2 0 0 1 2 1 1 1 2 1 1
#> 
#> $mmB[[9]]
#>   [1] 1 2 1 2 2 2 1 1 1 2 0 0 1 2 1 1 2 1 1 1 1 1 1 1 1 1 0 0 2 1 1 1 2 1 2 2 1
#>  [38] 0 1 1 2 1 1 1 2 2 2 1 1 2 2 2 0 0 1 1 1 1 2 1 1 1 2 0 2 2 2 2 1 2 1 1 0 0
#>  [75] 1 1 2 1 2 2 1 1 2 2 1 1 1 2 2 2 2 1 1 1 2 2 1 1 1 1 2 0 2 2 2 2 1 1 2 1 1
#> [112] 2 2 2 1 2 1 1 2 1 1 1 1 2 1 0 1 0 1 1 1 2 0 2 1 0 1 2 1 2
#> 
#> $mmB[[10]]
#>   [1] 1 1 1 0 1 1 2 2 2 1 2 0 1 1 2 2 2 1 2 2 1 1 0 0 1 2 2 1 1 1 2 1 2 2 1 1 0
#>  [38] 1 1 1 1 1 2 1 1 2 1 1 0 0 1 1 2 1 2 1 1 1 1 2 1 2 2 1 1 1 0 1 1 1 0 0 2 2
#>  [75] 2 1 1 0 1 2 1 1 1 1 1 2 1 1 1 2 2 1 2 1 1 1 2 2 1 2 1 2 2 2 2 2 1 2 2 2 2
#> [112] 2 2 1 1 0 1 1 1 1 1 1 1 1 2 2 1 1 2 1 1 1 1 1 1 1 1 2 1 2
#> 
#> 
#> $mmDR
#> $mmDR[[1]]
#>   [1] 0 1 1 0 1 2 1 2 2 2 1 2 2 1 0 0 1 1 0 2 2 1 1 2 2 2 1 2 1 2 2 1 2 1 2 2 2
#>  [38] 1 1 1 0 0 2 1 2 2 1 2 1 2 2 2 2 1 2 2 1 1 1 2 1 2 1 1 0 1 1 2 1 1 1 1 1 1
#>  [75] 1 1 1 0 1 1 1 0 1 2 2 1 1 1 2 2 2 1 1 2 1 0 1 2 1 1 2 2 1 2 1 2 2 0 1 1 2
#> [112] 2 2 2 1 2 2 0 1 2 1 2 1 1 2 0 2 2 1 1 0 2 2 1 2 1 0 0 0 1
#> 
#> $mmDR[[2]]
#>   [1] 1 2 2 2 2 0 2 2 0 1 2 1 1 1 2 0 1 2 1 2 1 1 0 1 0 1 1 1 1 1 2 1 1 1 1 1 0
#>  [38] 0 1 2 2 0 0 0 2 2 2 1 2 1 2 1 1 2 1 1 2 2 1 2 0 0 1 1 1 2 0 0 2 1 1 1 1 0
#>  [75] 2 0 0 2 1 1 1 1 1 1 1 2 1 2 1 2 2 2 2 1 1 2 2 2 2 1 1 1 2 2 1 1 1 2 1 1 1
#> [112] 1 2 2 2 1 2 2 1 1 2 2 1 1 1 2 1 0 2 1 2 2 1 1 2 1 1 1 1 2
#> 
#> $mmDR[[3]]
#>   [1] 1 1 1 2 1 1 1 1 1 2 1 2 0 0 1 2 2 0 0 0 1 1 0 2 1 1 2 2 1 1 2 2 2 2 1 1 1
#>  [38] 2 1 1 1 1 1 1 1 2 0 0 2 1 1 1 2 1 1 1 1 2 1 1 1 1 1 1 2 2 1 2 2 1 2 1 2 1
#>  [75] 1 1 0 2 2 0 1 1 1 1 1 2 1 1 0 2 1 2 1 1 1 2 2 2 2 2 2 2 2 2 1 1 1 1 1 1 1
#> [112] 2 1 2 2 1 1 1 1 2 2 1 2 1 1 1 1 0 2 1 2 2 2 1 1 1 1 1 0 0
#> 
#> $mmDR[[4]]
#>   [1] 2 1 1 1 1 2 2 1 1 1 1 2 2 1 1 2 2 1 2 2 1 1 2 1 1 0 1 1 1 2 1 1 1 2 1 1 1
#>  [38] 1 1 1 1 1 2 0 2 2 1 1 2 1 1 2 2 1 1 2 2 1 2 1 2 1 2 1 2 1 2 2 1 2 1 1 1 2
#>  [75] 0 1 1 1 1 2 2 1 2 2 1 2 1 0 2 1 1 1 1 1 2 1 1 2 1 1 1 0 1 1 1 1 1 1 0 2 1
#> [112] 1 1 1 2 2 0 0 1 0 2 0 1 2 1 0 1 1 2 1 2 2 0 0 2 1 1 1 1 0
#> 
#> $mmDR[[5]]
#>   [1] 1 1 2 1 1 1 2 0 1 2 1 1 0 0 1 1 1 1 1 1 1 2 0 1 2 1 1 1 1 1 1 1 2 2 1 1 2
#>  [38] 2 1 1 0 0 0 0 0 2 1 1 1 1 1 0 1 1 2 2 0 1 1 0 1 1 1 2 2 2 2 2 1 1 1 1 1 1
#>  [75] 1 2 2 0 2 0 2 2 1 1 1 1 1 1 1 2 2 2 2 1 2 1 1 2 1 1 1 1 2 2 2 1 2 2 2 2 1
#> [112] 2 2 1 1 2 1 1 1 1 2 1 2 2 2 0 1 1 2 1 0 2 0 0 1 0 1 2 1 1
#> 
#> $mmDR[[6]]
#>   [1] 1 2 2 2 1 1 1 1 1 2 2 2 1 2 2 1 1 1 0 0 1 2 1 0 1 2 1 1 1 1 2 0 0 1 2 1 1
#>  [38] 1 2 2 1 1 2 2 0 0 1 1 1 1 1 1 1 0 1 1 1 1 0 2 1 2 1 1 2 2 1 1 2 0 1 2 1 1
#>  [75] 1 2 1 1 1 1 0 0 1 1 2 2 2 2 1 1 2 1 1 0 0 1 2 1 2 2 0 2 2 1 1 2 1 1 1 1 2
#> [112] 1 0 1 2 2 2 2 2 2 1 2 1 2 2 2 1 0 0 1 2 0 2 1 1 0 1 0 2 2
#> 
#> $mmDR[[7]]
#>   [1] 2 0 2 2 0 1 2 2 2 0 2 2 2 0 2 1 1 1 1 1 1 1 2 1 1 1 2 2 1 2 0 1 1 2 1 1 2
#>  [38] 1 1 2 1 2 1 2 1 0 0 0 1 1 1 1 1 1 1 1 0 1 1 1 1 1 2 1 1 2 2 2 1 1 2 1 2 0
#>  [75] 1 1 1 1 2 2 1 1 2 1 2 1 1 2 2 1 1 1 2 1 1 2 1 2 1 2 0 0 1 2 1 2 0 0 1 1 2
#> [112] 0 1 2 1 1 2 1 1 1 1 1 1 2 2 2 1 1 2 1 2 1 1 0 0 0 1 2 1 0
#> 
#> $mmDR[[8]]
#>   [1] 1 1 2 2 1 2 2 1 1 2 2 1 0 1 1 2 2 1 1 2 0 0 2 1 1 1 2 1 1 2 1 1 0 0 1 1 2
#>  [38] 2 1 1 1 2 1 0 2 0 0 1 2 2 1 2 1 1 1 1 2 2 2 0 1 2 0 1 1 2 1 1 1 1 2 1 1 1
#>  [75] 1 1 1 1 2 1 2 2 1 1 2 2 1 2 2 2 1 1 2 2 1 1 2 1 2 2 2 2 1 0 2 2 1 1 1 1 1
#> [112] 2 1 1 0 0 1 1 1 0 1 1 2 2 2 0 1 1 2 2 0 0 1 2 1 1 1 2 1 1
#> 
#> $mmDR[[9]]
#>   [1] 1 2 1 2 2 2 1 1 1 2 0 0 1 2 1 1 2 1 1 1 1 1 1 1 1 1 0 0 2 1 1 1 2 1 2 2 1
#>  [38] 0 1 1 2 1 1 1 2 2 2 1 1 2 2 2 0 0 1 1 1 1 2 1 1 1 2 0 2 2 2 2 1 2 1 1 0 0
#>  [75] 1 1 2 1 2 2 1 1 2 2 1 1 1 2 2 2 2 1 1 1 2 2 1 1 1 1 2 0 2 2 2 2 1 1 2 1 1
#> [112] 2 2 2 1 2 1 1 2 1 1 1 1 2 1 0 1 0 1 1 1 2 0 2 1 0 1 2 1 2
#> 
#> $mmDR[[10]]
#>   [1] 1 1 1 0 1 1 2 2 2 1 2 0 1 1 2 2 2 1 2 2 1 1 0 0 1 2 2 1 1 1 2 1 2 2 1 1 0
#>  [38] 1 1 1 1 1 2 1 1 2 1 1 0 0 1 1 2 1 2 1 1 1 1 2 1 2 2 1 1 1 0 1 1 1 0 0 2 2
#>  [75] 2 1 1 0 1 2 1 1 1 1 1 2 1 1 1 2 2 1 2 1 1 1 2 2 1 2 1 2 2 2 2 2 1 2 2 2 2
#> [112] 2 2 1 1 0 1 1 1 1 1 1 1 1 2 2 1 1 2 1 1 1 1 1 1 1 1 2 1 2
#> 
#> 
#> $ABO
#> $ABO[[1]]
#>   [1] 0 1 1 0 1 2 1 2 2 2 1 2 2 1 0 0 1 1 0 2 2 1 1 2 2 2 1 2 1 2 2 1 2 1 2 2 2
#>  [38] 1 1 1 0 0 2 1 2 2 1 2 1 2 2 2 2 1 2 2 1 1 1 2 1 2 1 1 0 1 1 2 1 1 1 1 1 1
#>  [75] 1 1 1 0 1 1 1 0 1 2 2 1 1 1 2 2 2 1 1 2 1 0 1 2 1 1 2 2 1 2 1 2 2 0 1 1 2
#> [112] 2 2 2 1 2 2 0 1 2 1 2 1 1 2 0 2 2 1 1 0 2 2 1 2 1 0 0 0 1
#> 
#> $ABO[[2]]
#>   [1] 1 2 2 2 2 0 2 2 0 1 2 1 1 1 2 0 1 2 1 2 1 1 0 1 0 1 1 1 1 1 2 1 1 1 1 1 0
#>  [38] 0 1 2 2 0 0 0 2 2 2 1 2 1 2 1 1 2 1 1 2 2 1 2 0 0 1 1 1 2 0 0 2 1 1 1 1 0
#>  [75] 2 0 0 2 1 1 1 1 1 1 1 2 1 2 1 2 2 2 2 1 1 2 2 2 2 1 1 1 2 2 1 1 1 2 1 1 1
#> [112] 1 2 2 2 1 2 2 1 1 2 2 1 1 1 2 1 0 2 1 2 2 1 1 2 1 1 1 1 2
#> 
#> $ABO[[3]]
#>   [1] 1 1 1 2 1 1 1 1 1 2 1 2 0 0 1 2 2 0 0 0 1 1 0 2 1 1 2 2 1 1 2 2 2 2 1 1 1
#>  [38] 2 1 1 1 1 1 1 1 2 0 0 2 1 1 1 2 1 1 1 1 2 1 1 1 1 1 1 2 2 1 2 2 1 2 1 2 1
#>  [75] 1 1 0 2 2 0 1 1 1 1 1 2 1 1 0 2 1 2 1 1 1 2 2 2 2 2 2 2 2 2 1 1 1 1 1 1 1
#> [112] 2 1 2 2 1 1 1 1 2 2 1 2 1 1 1 1 0 2 1 2 2 2 1 1 1 1 1 0 0
#> 
#> $ABO[[4]]
#>   [1] 2 1 1 1 1 2 2 1 1 1 1 2 2 1 1 2 2 1 2 2 1 1 2 1 1 0 1 1 1 2 1 1 1 2 1 1 1
#>  [38] 1 1 1 1 1 2 0 2 2 1 1 2 1 1 2 2 1 1 2 2 1 2 1 2 1 2 1 2 1 2 2 1 2 1 1 1 2
#>  [75] 0 1 1 1 1 2 2 1 2 2 1 2 1 0 2 1 1 1 1 1 2 1 1 2 1 1 1 0 1 1 1 1 1 1 0 2 1
#> [112] 1 1 1 2 2 0 0 1 0 2 0 1 2 1 0 1 1 2 1 2 2 0 0 2 1 1 1 1 0
#> 
#> $ABO[[5]]
#>   [1] 1 1 2 1 1 1 2 0 1 2 1 1 0 0 1 1 1 1 1 1 1 2 0 1 2 1 1 1 1 1 1 1 2 2 1 1 2
#>  [38] 2 1 1 0 0 0 0 0 2 1 1 1 1 1 0 1 1 2 2 0 1 1 0 1 1 1 2 2 2 2 2 1 1 1 1 1 1
#>  [75] 1 2 2 0 2 0 2 2 1 1 1 1 1 1 1 2 2 2 2 1 2 1 1 2 1 1 1 1 2 2 2 1 2 2 2 2 1
#> [112] 2 2 1 1 2 1 1 1 1 2 1 2 2 2 0 1 1 2 1 0 2 0 0 1 0 1 2 1 1
#> 
#> $ABO[[6]]
#>   [1] 1 2 2 2 1 1 1 1 1 2 2 2 1 2 2 1 1 1 0 0 1 2 1 0 1 2 1 1 1 1 2 0 0 1 2 1 1
#>  [38] 1 2 2 1 1 2 2 0 0 1 1 1 1 1 1 1 0 1 1 1 1 0 2 1 2 1 1 2 2 1 1 2 0 1 2 1 1
#>  [75] 1 2 1 1 1 1 0 0 1 1 2 2 2 2 1 1 2 1 1 0 0 1 2 1 2 2 0 2 2 1 1 2 1 1 1 1 2
#> [112] 1 0 1 2 2 2 2 2 2 1 2 1 2 2 2 1 0 0 1 2 0 2 1 1 0 1 0 2 2
#> 
#> $ABO[[7]]
#>   [1] 2 0 2 2 0 1 2 2 2 0 2 2 2 0 2 1 1 1 1 1 1 1 2 1 1 1 2 2 1 2 0 1 1 2 1 1 2
#>  [38] 1 1 2 1 2 1 2 1 0 0 0 1 1 1 1 1 1 1 1 0 1 1 1 1 1 2 1 1 2 2 2 1 1 2 1 2 0
#>  [75] 1 1 1 1 2 2 1 1 2 1 2 1 1 2 2 1 1 1 2 1 1 2 1 2 1 2 0 0 1 2 1 2 0 0 1 1 2
#> [112] 0 1 2 1 1 2 1 1 1 1 1 1 2 2 2 1 1 2 1 2 1 1 0 0 0 1 2 1 0
#> 
#> $ABO[[8]]
#>   [1] 1 1 2 2 1 2 2 1 1 2 2 1 0 1 1 2 2 1 1 2 0 0 2 1 1 1 2 1 1 2 1 1 0 0 1 1 2
#>  [38] 2 1 1 1 2 1 0 2 0 0 1 2 2 1 2 1 1 1 1 2 2 2 0 1 2 0 1 1 2 1 1 1 1 2 1 1 1
#>  [75] 1 1 1 1 2 1 2 2 1 1 2 2 1 2 2 2 1 1 2 2 1 1 2 1 2 2 2 2 1 0 2 2 1 1 1 1 1
#> [112] 2 1 1 0 0 1 1 1 0 1 1 2 2 2 0 1 1 2 2 0 0 1 2 1 1 1 2 1 1
#> 
#> $ABO[[9]]
#>   [1] 1 2 1 2 2 2 1 1 1 2 0 0 1 2 1 1 2 1 1 1 1 1 1 1 1 1 0 0 2 1 1 1 2 1 2 2 1
#>  [38] 0 1 1 2 1 1 1 2 2 2 1 1 2 2 2 0 0 1 1 1 1 2 1 1 1 2 0 2 2 2 2 1 2 1 1 0 0
#>  [75] 1 1 2 1 2 2 1 1 2 2 1 1 1 2 2 2 2 1 1 1 2 2 1 1 1 1 2 0 2 2 2 2 1 1 2 1 1
#> [112] 2 2 2 1 2 1 1 2 1 1 1 1 2 1 0 1 0 1 1 1 2 0 2 1 0 1 2 1 2
#> 
#> $ABO[[10]]
#>   [1] 1 1 1 0 1 1 2 2 2 1 2 0 1 1 2 2 2 1 2 2 1 1 0 0 1 2 2 1 1 1 2 1 2 2 1 1 0
#>  [38] 1 1 1 1 1 2 1 1 2 1 1 0 0 1 1 2 1 2 1 1 1 1 2 1 2 2 1 1 1 0 1 1 1 0 0 2 2
#>  [75] 2 1 1 0 1 2 1 1 1 1 1 2 1 1 1 2 2 1 2 1 1 1 2 2 1 2 1 2 2 2 2 2 1 2 2 2 2
#> [112] 2 2 1 1 0 1 1 1 1 1 1 1 1 2 2 1 1 2 1 1 1 1 1 1 1 1 2 1 2
#> 
#> 
#> $HI
#> $HI[[1]]
#>   [1] 0 1 1 0 1 2 1 2 2 2 1 2 2 1 0 0 1 1 0 2 2 1 1 2 2 2 1 2 1 2 2 1 2 1 2 2 2
#>  [38] 1 1 1 0 0 2 1 2 2 1 2 1 2 2 2 2 1 2 2 1 1 1 2 1 2 1 1 0 1 1 2 1 1 1 1 1 1
#>  [75] 1 1 1 0 1 1 1 0 1 2 2 1 1 1 2 2 2 1 1 2 1 0 1 2 1 1 2 2 1 2 1 2 2 0 1 1 2
#> [112] 2 2 2 1 2 2 0 1 2 1 2 1 1 2 0 2 2 1 1 0 2 2 1 2 1 0 0 0 1
#> 
#> $HI[[2]]
#>   [1] 1 2 2 2 2 0 2 2 0 1 2 1 1 1 2 0 1 2 1 2 1 1 0 1 0 1 1 1 1 1 2 1 1 1 1 1 0
#>  [38] 0 1 2 2 0 0 0 2 2 2 1 2 1 2 1 1 2 1 1 2 2 1 2 0 0 1 1 1 2 0 0 2 1 1 1 1 0
#>  [75] 2 0 0 2 1 1 1 1 1 1 1 2 1 2 1 2 2 2 2 1 1 2 2 2 2 1 1 1 2 2 1 1 1 2 1 1 1
#> [112] 1 2 2 2 1 2 2 1 1 2 2 1 1 1 2 1 0 2 1 2 2 1 1 2 1 1 1 1 2
#> 
#> $HI[[3]]
#>   [1] 1 1 1 2 1 1 1 1 1 2 1 2 0 0 1 2 2 0 0 0 1 1 0 2 1 1 2 2 1 1 2 2 2 2 1 1 1
#>  [38] 2 1 1 1 1 1 1 1 2 0 0 2 1 1 1 2 1 1 1 1 2 1 1 1 1 1 1 2 2 1 2 2 1 2 1 2 1
#>  [75] 1 1 0 2 2 0 1 1 1 1 1 2 1 1 0 2 1 2 1 1 1 2 2 2 2 2 2 2 2 2 1 1 1 1 1 1 1
#> [112] 2 1 2 2 1 1 1 1 2 2 1 2 1 1 1 1 0 2 1 2 2 2 1 1 1 1 1 0 0
#> 
#> $HI[[4]]
#>   [1] 2 1 1 1 1 2 2 1 1 1 1 2 2 1 1 2 2 1 2 2 1 1 2 1 1 0 1 1 1 2 1 1 1 2 1 1 1
#>  [38] 1 1 1 1 1 2 0 2 2 1 1 2 1 1 2 2 1 1 2 2 1 2 1 2 1 2 1 2 1 2 2 1 2 1 1 1 2
#>  [75] 0 1 1 1 1 2 2 1 2 2 1 2 1 0 2 1 1 1 1 1 2 1 1 2 1 1 1 0 1 1 1 1 1 1 0 2 1
#> [112] 1 1 1 2 2 0 0 1 0 2 0 1 2 1 0 1 1 2 1 2 2 0 0 2 1 1 1 1 0
#> 
#> $HI[[5]]
#>   [1] 1 1 2 1 1 1 2 0 1 2 1 1 0 0 1 1 1 1 1 1 1 2 0 1 2 1 1 1 1 1 1 1 2 2 1 1 2
#>  [38] 2 1 1 0 0 0 0 0 2 1 1 1 1 1 0 1 1 2 2 0 1 1 0 1 1 1 2 2 2 2 2 1 1 1 1 1 1
#>  [75] 1 2 2 0 2 0 2 2 1 1 1 1 1 1 1 2 2 2 2 1 2 1 1 2 1 1 1 1 2 2 2 1 2 2 2 2 1
#> [112] 2 2 1 1 2 1 1 1 1 2 1 2 2 2 0 1 1 2 1 0 2 0 0 1 0 1 2 1 1
#> 
#> $HI[[6]]
#>   [1] 1 2 2 2 1 1 1 1 1 2 2 2 1 2 2 1 1 1 0 0 1 2 1 0 1 2 1 1 1 1 2 0 0 1 2 1 1
#>  [38] 1 2 2 1 1 2 2 0 0 1 1 1 1 1 1 1 0 1 1 1 1 0 2 1 2 1 1 2 2 1 1 2 0 1 2 1 1
#>  [75] 1 2 1 1 1 1 0 0 1 1 2 2 2 2 1 1 2 1 1 0 0 1 2 1 2 2 0 2 2 1 1 2 1 1 1 1 2
#> [112] 1 0 1 2 2 2 2 2 2 1 2 1 2 2 2 1 0 0 1 2 0 2 1 1 0 1 0 2 2
#> 
#> $HI[[7]]
#>   [1] 2 0 2 2 0 1 2 2 2 0 2 2 2 0 2 1 1 1 1 1 1 1 2 1 1 1 2 2 1 2 0 1 1 2 1 1 2
#>  [38] 1 1 2 1 2 1 2 1 0 0 0 1 1 1 1 1 1 1 1 0 1 1 1 1 1 2 1 1 2 2 2 1 1 2 1 2 0
#>  [75] 1 1 1 1 2 2 1 1 2 1 2 1 1 2 2 1 1 1 2 1 1 2 1 2 1 2 0 0 1 2 1 2 0 0 1 1 2
#> [112] 0 1 2 1 1 2 1 1 1 1 1 1 2 2 2 1 1 2 1 2 1 1 0 0 0 1 2 1 0
#> 
#> $HI[[8]]
#>   [1] 1 1 2 2 1 2 2 1 1 2 2 1 0 1 1 2 2 1 1 2 0 0 2 1 1 1 2 1 1 2 1 1 0 0 1 1 2
#>  [38] 2 1 1 1 2 1 0 2 0 0 1 2 2 1 2 1 1 1 1 2 2 2 0 1 2 0 1 1 2 1 1 1 1 2 1 1 1
#>  [75] 1 1 1 1 2 1 2 2 1 1 2 2 1 2 2 2 1 1 2 2 1 1 2 1 2 2 2 2 1 0 2 2 1 1 1 1 1
#> [112] 2 1 1 0 0 1 1 1 0 1 1 2 2 2 0 1 1 2 2 0 0 1 2 1 1 1 2 1 1
#> 
#> $HI[[9]]
#>   [1] 1 2 1 2 2 2 1 1 1 2 0 0 1 2 1 1 2 1 1 1 1 1 1 1 1 1 0 0 2 1 1 1 2 1 2 2 1
#>  [38] 0 1 1 2 1 1 1 2 2 2 1 1 2 2 2 0 0 1 1 1 1 2 1 1 1 2 0 2 2 2 2 1 2 1 1 0 0
#>  [75] 1 1 2 1 2 2 1 1 2 2 1 1 1 2 2 2 2 1 1 1 2 2 1 1 1 1 2 0 2 2 2 2 1 1 2 1 1
#> [112] 2 2 2 1 2 1 1 2 1 1 1 1 2 1 0 1 0 1 1 1 2 0 2 1 0 1 2 1 2
#> 
#> $HI[[10]]
#>   [1] 1 1 1 0 1 1 2 2 2 1 2 0 1 1 2 2 2 1 2 2 1 1 0 0 1 2 2 1 1 1 2 1 2 2 1 1 0
#>  [38] 1 1 1 1 1 2 1 1 2 1 1 0 0 1 1 2 1 2 1 1 1 1 2 1 2 2 1 1 1 0 1 1 1 0 0 2 2
#>  [75] 2 1 1 0 1 2 1 1 1 1 1 2 1 1 1 2 2 1 2 1 1 1 2 2 1 2 1 2 2 2 2 2 1 2 2 2 2
#> [112] 2 2 1 1 0 1 1 1 1 1 1 1 1 2 2 1 1 2 1 1 1 1 1 1 1 1 2 1 2
#> 
#> 
#> $color
#> $color[[1]]
#>   [1] 0 1 1 0 1 2 1 2 2 2 1 2 2 1 0 0 1 1 0 2 2 1 1 2 2 2 1 2 1 2 2 1 2 1 2 2 2
#>  [38] 1 1 1 0 0 2 1 2 2 1 2 1 2 2 2 2 1 2 2 1 1 1 2 1 2 1 1 0 1 1 2 1 1 1 1 1 1
#>  [75] 1 1 1 0 1 1 1 0 1 2 2 1 1 1 2 2 2 1 1 2 1 0 1 2 1 1 2 2 1 2 1 2 2 0 1 1 2
#> [112] 2 2 2 1 2 2 0 1 2 1 2 1 1 2 0 2 2 1 1 0 2 2 1 2 1 0 0 0 1
#> 
#> $color[[2]]
#>   [1] 1 2 2 2 2 0 2 2 0 1 2 1 1 1 2 0 1 2 1 2 1 1 0 1 0 1 1 1 1 1 2 1 1 1 1 1 0
#>  [38] 0 1 2 2 0 0 0 2 2 2 1 2 1 2 1 1 2 1 1 2 2 1 2 0 0 1 1 1 2 0 0 2 1 1 1 1 0
#>  [75] 2 0 0 2 1 1 1 1 1 1 1 2 1 2 1 2 2 2 2 1 1 2 2 2 2 1 1 1 2 2 1 1 1 2 1 1 1
#> [112] 1 2 2 2 1 2 2 1 1 2 2 1 1 1 2 1 0 2 1 2 2 1 1 2 1 1 1 1 2
#> 
#> $color[[3]]
#>   [1] 1 1 1 2 1 1 1 1 1 2 1 2 0 0 1 2 2 0 0 0 1 1 0 2 1 1 2 2 1 1 2 2 2 2 1 1 1
#>  [38] 2 1 1 1 1 1 1 1 2 0 0 2 1 1 1 2 1 1 1 1 2 1 1 1 1 1 1 2 2 1 2 2 1 2 1 2 1
#>  [75] 1 1 0 2 2 0 1 1 1 1 1 2 1 1 0 2 1 2 1 1 1 2 2 2 2 2 2 2 2 2 1 1 1 1 1 1 1
#> [112] 2 1 2 2 1 1 1 1 2 2 1 2 1 1 1 1 0 2 1 2 2 2 1 1 1 1 1 0 0
#> 
#> $color[[4]]
#>   [1] 2 1 1 1 1 2 2 1 1 1 1 2 2 1 1 2 2 1 2 2 1 1 2 1 1 0 1 1 1 2 1 1 1 2 1 1 1
#>  [38] 1 1 1 1 1 2 0 2 2 1 1 2 1 1 2 2 1 1 2 2 1 2 1 2 1 2 1 2 1 2 2 1 2 1 1 1 2
#>  [75] 0 1 1 1 1 2 2 1 2 2 1 2 1 0 2 1 1 1 1 1 2 1 1 2 1 1 1 0 1 1 1 1 1 1 0 2 1
#> [112] 1 1 1 2 2 0 0 1 0 2 0 1 2 1 0 1 1 2 1 2 2 0 0 2 1 1 1 1 0
#> 
#> $color[[5]]
#>   [1] 1 1 2 1 1 1 2 0 1 2 1 1 0 0 1 1 1 1 1 1 1 2 0 1 2 1 1 1 1 1 1 1 2 2 1 1 2
#>  [38] 2 1 1 0 0 0 0 0 2 1 1 1 1 1 0 1 1 2 2 0 1 1 0 1 1 1 2 2 2 2 2 1 1 1 1 1 1
#>  [75] 1 2 2 0 2 0 2 2 1 1 1 1 1 1 1 2 2 2 2 1 2 1 1 2 1 1 1 1 2 2 2 1 2 2 2 2 1
#> [112] 2 2 1 1 2 1 1 1 1 2 1 2 2 2 0 1 1 2 1 0 2 0 0 1 0 1 2 1 1
#> 
#> $color[[6]]
#>   [1] 1 2 2 2 1 1 1 1 1 2 2 2 1 2 2 1 1 1 0 0 1 2 1 0 1 2 1 1 1 1 2 0 0 1 2 1 1
#>  [38] 1 2 2 1 1 2 2 0 0 1 1 1 1 1 1 1 0 1 1 1 1 0 2 1 2 1 1 2 2 1 1 2 0 1 2 1 1
#>  [75] 1 2 1 1 1 1 0 0 1 1 2 2 2 2 1 1 2 1 1 0 0 1 2 1 2 2 0 2 2 1 1 2 1 1 1 1 2
#> [112] 1 0 1 2 2 2 2 2 2 1 2 1 2 2 2 1 0 0 1 2 0 2 1 1 0 1 0 2 2
#> 
#> $color[[7]]
#>   [1] 2 0 2 2 0 1 2 2 2 0 2 2 2 0 2 1 1 1 1 1 1 1 2 1 1 1 2 2 1 2 0 1 1 2 1 1 2
#>  [38] 1 1 2 1 2 1 2 1 0 0 0 1 1 1 1 1 1 1 1 0 1 1 1 1 1 2 1 1 2 2 2 1 1 2 1 2 0
#>  [75] 1 1 1 1 2 2 1 1 2 1 2 1 1 2 2 1 1 1 2 1 1 2 1 2 1 2 0 0 1 2 1 2 0 0 1 1 2
#> [112] 0 1 2 1 1 2 1 1 1 1 1 1 2 2 2 1 1 2 1 2 1 1 0 0 0 1 2 1 0
#> 
#> $color[[8]]
#>   [1] 1 1 2 2 1 2 2 1 1 2 2 1 0 1 1 2 2 1 1 2 0 0 2 1 1 1 2 1 1 2 1 1 0 0 1 1 2
#>  [38] 2 1 1 1 2 1 0 2 0 0 1 2 2 1 2 1 1 1 1 2 2 2 0 1 2 0 1 1 2 1 1 1 1 2 1 1 1
#>  [75] 1 1 1 1 2 1 2 2 1 1 2 2 1 2 2 2 1 1 2 2 1 1 2 1 2 2 2 2 1 0 2 2 1 1 1 1 1
#> [112] 2 1 1 0 0 1 1 1 0 1 1 2 2 2 0 1 1 2 2 0 0 1 2 1 1 1 2 1 1
#> 
#> $color[[9]]
#>   [1] 1 2 1 2 2 2 1 1 1 2 0 0 1 2 1 1 2 1 1 1 1 1 1 1 1 1 0 0 2 1 1 1 2 1 2 2 1
#>  [38] 0 1 1 2 1 1 1 2 2 2 1 1 2 2 2 0 0 1 1 1 1 2 1 1 1 2 0 2 2 2 2 1 2 1 1 0 0
#>  [75] 1 1 2 1 2 2 1 1 2 2 1 1 1 2 2 2 2 1 1 1 2 2 1 1 1 1 2 0 2 2 2 2 1 1 2 1 1
#> [112] 2 2 2 1 2 1 1 2 1 1 1 1 2 1 0 1 0 1 1 1 2 0 2 1 0 1 2 1 2
#> 
#> $color[[10]]
#>   [1] 1 1 1 0 1 1 2 2 2 1 2 0 1 1 2 2 2 1 2 2 1 1 0 0 1 2 2 1 1 1 2 1 2 2 1 1 0
#>  [38] 1 1 1 1 1 2 1 1 2 1 1 0 0 1 1 2 1 2 1 1 1 1 2 1 2 2 1 1 1 0 1 1 1 0 0 2 2
#>  [75] 2 1 1 0 1 2 1 1 1 1 1 2 1 1 1 2 2 1 2 1 1 1 2 2 1 2 1 2 2 2 2 2 1 2 2 2 2
#> [112] 2 2 1 1 0 1 1 1 1 1 1 1 1 2 2 1 1 2 1 1 1 1 1 1 1 1 2 1 2
#> 
#> 
#> $SP
#> $SP[[1]]
#>   [1] 0 1 1 0 1 2 1 2 2 2 1 2 2 1 0 0 1 1 0 2 2 1 1 2 2 2 1 2 1 2 2 1 2 1 2 2 2
#>  [38] 1 1 1 0 0 2 1 2 2 1 2 1 2 2 2 2 1 2 2 1 1 1 2 1 2 1 1 0 1 1 2 1 1 1 1 1 1
#>  [75] 1 1 1 0 1 1 1 0 1 2 2 1 1 1 2 2 2 1 1 2 1 0 1 2 1 1 2 2 1 2 1 2 2 0 1 1 2
#> [112] 2 2 2 1 2 2 0 1 2 1 2 1 1 2 0 2 2 1 1 0 2 2 1 2 1 0 0 0 1
#> 
#> $SP[[2]]
#>   [1] 1 2 2 2 2 0 2 2 0 1 2 1 1 1 2 0 1 2 1 2 1 1 0 1 0 1 1 1 1 1 2 1 1 1 1 1 0
#>  [38] 0 1 2 2 0 0 0 2 2 2 1 2 1 2 1 1 2 1 1 2 2 1 2 0 0 1 1 1 2 0 0 2 1 1 1 1 0
#>  [75] 2 0 0 2 1 1 1 1 1 1 1 2 1 2 1 2 2 2 2 1 1 2 2 2 2 1 1 1 2 2 1 1 1 2 1 1 1
#> [112] 1 2 2 2 1 2 2 1 1 2 2 1 1 1 2 1 0 2 1 2 2 1 1 2 1 1 1 1 2
#> 
#> $SP[[3]]
#>   [1] 1 1 1 2 1 1 1 1 1 2 1 2 0 0 1 2 2 0 0 0 1 1 0 2 1 1 2 2 1 1 2 2 2 2 1 1 1
#>  [38] 2 1 1 1 1 1 1 1 2 0 0 2 1 1 1 2 1 1 1 1 2 1 1 1 1 1 1 2 2 1 2 2 1 2 1 2 1
#>  [75] 1 1 0 2 2 0 1 1 1 1 1 2 1 1 0 2 1 2 1 1 1 2 2 2 2 2 2 2 2 2 1 1 1 1 1 1 1
#> [112] 2 1 2 2 1 1 1 1 2 2 1 2 1 1 1 1 0 2 1 2 2 2 1 1 1 1 1 0 0
#> 
#> $SP[[4]]
#>   [1] 2 1 1 1 1 2 2 1 1 1 1 2 2 1 1 2 2 1 2 2 1 1 2 1 1 0 1 1 1 2 1 1 1 2 1 1 1
#>  [38] 1 1 1 1 1 2 0 2 2 1 1 2 1 1 2 2 1 1 2 2 1 2 1 2 1 2 1 2 1 2 2 1 2 1 1 1 2
#>  [75] 0 1 1 1 1 2 2 1 2 2 1 2 1 0 2 1 1 1 1 1 2 1 1 2 1 1 1 0 1 1 1 1 1 1 0 2 1
#> [112] 1 1 1 2 2 0 0 1 0 2 0 1 2 1 0 1 1 2 1 2 2 0 0 2 1 1 1 1 0
#> 
#> $SP[[5]]
#>   [1] 1 1 2 1 1 1 2 0 1 2 1 1 0 0 1 1 1 1 1 1 1 2 0 1 2 1 1 1 1 1 1 1 2 2 1 1 2
#>  [38] 2 1 1 0 0 0 0 0 2 1 1 1 1 1 0 1 1 2 2 0 1 1 0 1 1 1 2 2 2 2 2 1 1 1 1 1 1
#>  [75] 1 2 2 0 2 0 2 2 1 1 1 1 1 1 1 2 2 2 2 1 2 1 1 2 1 1 1 1 2 2 2 1 2 2 2 2 1
#> [112] 2 2 1 1 2 1 1 1 1 2 1 2 2 2 0 1 1 2 1 0 2 0 0 1 0 1 2 1 1
#> 
#> $SP[[6]]
#>   [1] 1 2 2 2 1 1 1 1 1 2 2 2 1 2 2 1 1 1 0 0 1 2 1 0 1 2 1 1 1 1 2 0 0 1 2 1 1
#>  [38] 1 2 2 1 1 2 2 0 0 1 1 1 1 1 1 1 0 1 1 1 1 0 2 1 2 1 1 2 2 1 1 2 0 1 2 1 1
#>  [75] 1 2 1 1 1 1 0 0 1 1 2 2 2 2 1 1 2 1 1 0 0 1 2 1 2 2 0 2 2 1 1 2 1 1 1 1 2
#> [112] 1 0 1 2 2 2 2 2 2 1 2 1 2 2 2 1 0 0 1 2 0 2 1 1 0 1 0 2 2
#> 
#> $SP[[7]]
#>   [1] 2 0 2 2 0 1 2 2 2 0 2 2 2 0 2 1 1 1 1 1 1 1 2 1 1 1 2 2 1 2 0 1 1 2 1 1 2
#>  [38] 1 1 2 1 2 1 2 1 0 0 0 1 1 1 1 1 1 1 1 0 1 1 1 1 1 2 1 1 2 2 2 1 1 2 1 2 0
#>  [75] 1 1 1 1 2 2 1 1 2 1 2 1 1 2 2 1 1 1 2 1 1 2 1 2 1 2 0 0 1 2 1 2 0 0 1 1 2
#> [112] 0 1 2 1 1 2 1 1 1 1 1 1 2 2 2 1 1 2 1 2 1 1 0 0 0 1 2 1 0
#> 
#> $SP[[8]]
#>   [1] 1 1 2 2 1 2 2 1 1 2 2 1 0 1 1 2 2 1 1 2 0 0 2 1 1 1 2 1 1 2 1 1 0 0 1 1 2
#>  [38] 2 1 1 1 2 1 0 2 0 0 1 2 2 1 2 1 1 1 1 2 2 2 0 1 2 0 1 1 2 1 1 1 1 2 1 1 1
#>  [75] 1 1 1 1 2 1 2 2 1 1 2 2 1 2 2 2 1 1 2 2 1 1 2 1 2 2 2 2 1 0 2 2 1 1 1 1 1
#> [112] 2 1 1 0 0 1 1 1 0 1 1 2 2 2 0 1 1 2 2 0 0 1 2 1 1 1 2 1 1
#> 
#> $SP[[9]]
#>   [1] 1 2 1 2 2 2 1 1 1 2 0 0 1 2 1 1 2 1 1 1 1 1 1 1 1 1 0 0 2 1 1 1 2 1 2 2 1
#>  [38] 0 1 1 2 1 1 1 2 2 2 1 1 2 2 2 0 0 1 1 1 1 2 1 1 1 2 0 2 2 2 2 1 2 1 1 0 0
#>  [75] 1 1 2 1 2 2 1 1 2 2 1 1 1 2 2 2 2 1 1 1 2 2 1 1 1 1 2 0 2 2 2 2 1 1 2 1 1
#> [112] 2 2 2 1 2 1 1 2 1 1 1 1 2 1 0 1 0 1 1 1 2 0 2 1 0 1 2 1 2
#> 
#> $SP[[10]]
#>   [1] 1 1 1 0 1 1 2 2 2 1 2 0 1 1 2 2 2 1 2 2 1 1 0 0 1 2 2 1 1 1 2 1 2 2 1 1 0
#>  [38] 1 1 1 1 1 2 1 1 2 1 1 0 0 1 1 2 1 2 1 1 1 1 2 1 2 2 1 1 1 0 1 1 1 0 0 2 2
#>  [75] 2 1 1 0 1 2 1 1 1 1 1 2 1 1 1 2 2 1 2 1 1 1 2 2 1 2 1 2 2 2 2 2 1 2 2 2 2
#> [112] 2 2 1 1 0 1 1 1 1 1 1 1 1 2 2 1 1 2 1 1 1 1 1 1 1 1 2 1 2
#> 
#> 
# }